### 2022-06-29 #4523 @Kevin-Mattheus-Moerman
@masadcv thanks for this JOSS submission. I am the AEiC on call this week and here to help process the initial steps for this work. Before we proceed, can you check the following:

- [ ] Can you check those potentially missing DOIs? :point_up: You can call `@editorialbot generate pdf` here to update the paper and use `@editorialbot check references` to check the DOIs again. Thanks
