+++ paper/paper.bib
@@ -1,148 +1,151 @@
 @article{wang2018deepigeos,
-  title={DeepIGeoS: a deep interactive geodesic framework for medical image segmentation},
-  author={Wang, Guotai and Zuluaga, Maria A and Li, Wenqi and Pratt, Rosalind and Patel, Premal A and Aertsen, Michael and Doel, Tom and David, Anna L and Deprest, Jan and Ourselin, S{\'e}bastien and others},
-  journal={IEEE transactions on pattern analysis and machine intelligence},
-  volume={41},
-  number={7},
-  pages={1559--1572},
-  year={2018},
-  publisher={IEEE}
+  title     = {DeepIGeoS: a deep interactive geodesic framework for medical image segmentation},
+  author    = {Wang, Guotai and Zuluaga, Maria A and Li, Wenqi and Pratt, Rosalind and Patel, Premal A and Aertsen, Michael and Doel, Tom and David, Anna L and Deprest, Jan and Ourselin, S{\'e}bastien and others},
+  journal   = {IEEE transactions on pattern analysis and machine intelligence},
+  volume    = {41},
+  number    = {7},
+  pages     = {1559--1572},
+  year      = {2018},
+  publisher = {IEEE}
 }
 
 @inproceedings{criminisi2008geos,
-  title={Geos: Geodesic image segmentation},
-  author={Criminisi, Antonio and Sharp, Toby and Blake, Andrew},
-  booktitle={European Conference on Computer Vision},
-  pages={99--112},
-  year={2008},
-  organization={Springer}
+  title        = {Geos: Geodesic image segmentation},
+  author       = {Criminisi, Antonio and Sharp, Toby and Blake, Andrew},
+  booktitle    = {European Conference on Computer Vision},
+  pages        = {99--112},
+  year         = {2008},
+  organization = {Springer},
+  doi          = {10.1007/978-3-540-88682-2_9}
 }
 
 @article{weber2008parallel,
-  title={Parallel algorithms for approximation of distance maps on parametric surfaces},
-  author={Weber, Ofir and Devir, Yohai S and Bronstein, Alexander M and Bronstein, Michael M and Kimmel, Ron},
-  journal={ACM Transactions on Graphics (TOG)},
-  volume={27},
-  number={4},
-  pages={1--16},
-  year={2008},
-  publisher={ACM New York, NY, USA}
+  title     = {Parallel algorithms for approximation of distance maps on parametric surfaces},
+  author    = {Weber, Ofir and Devir, Yohai S and Bronstein, Alexander M and Bronstein, Michael M and Kimmel, Ron},
+  journal   = {ACM Transactions on Graphics (TOG)},
+  volume    = {27},
+  number    = {4},
+  pages     = {1--16},
+  year      = {2008},
+  publisher = {ACM New York, NY, USA},
+  doi       = {10.1145/1409625.1409626}
 }
 
 @misc{geodistk,
-  author = {Wang, Guotai},
-  title = {GeodisTK: Geodesic Distance Transform Toolkit for 2D and 3D Images},
-  year = {2020},
+  author    = {Wang, Guotai},
+  title     = {GeodisTK: Geodesic Distance Transform Toolkit for 2D and 3D Images},
+  year      = {2020},
   publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/taigw/GeodisTK}
+  journal   = {GitHub repository},
+  url       = {https://github.com/taigw/GeodisTK}
 }
 
 @misc{eucildeantdimpl,
-  author = {Seung-Lab, },
-  title = {Multi-Label Anisotropic 3D Euclidean Distance Transform (MLAEDT-3D)},
-  year = {2018},
+  author    = {Seung-Lab, },
[…]
