### 2022-06-30 #4532 @editorialbot
[REVIEW]: FastGeodis: Fast Generalised Geodesic Distance Transform
**Submitting author:** @masadcv (<a href="http://orcid.org/0000-0002-3672-2414">Muhammad Asad</a>)
**Repository:** https://github.com/masadcv/FastGeodis
**Branch with paper.md** (empty if default branch):
**Version:**  v1.0.0
