### 2022-07-16 292e0f0
adding multi-channel support to cuda kernels
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
