+++ FastGeodis/fastgeodis_cuda.cu
@@ -51,16 +51,6 @@ __device__ float l1distance_cuda(const float &in1, const float &in2)
     return abs(in1 - in2);
 }
 
-float l1distance_cuda(const float *in1, const float *in2, int size)
-{
-    float ret_sum = 0.0;
-    for (int c_i = 0; c_i < size; c_i++)
-    {
-        ret_sum += abs(in1[c_i] - in2[c_i]);
-    }
-    return ret_sum;
-}
-
 template <typename scalar_t>
 __global__ void geodesic_updown_single_row_pass_kernel(
     torch::PackedTensorAccessor32<scalar_t, 4, torch::RestrictPtrTraits> image_ptr, 
@@ -71,15 +61,15 @@ __global__ void geodesic_updown_single_row_pass_kernel(
     const int h
     )
 {
-    const float height = image_ptr.size(2);
-    const float width = image_ptr.size(3);
+    const int channel = image_ptr.size(1);
+    const int width = image_ptr.size(3);
 
     int kernelW = blockIdx.x * blockDim.x + threadIdx.x;
     
     // if outside, then skip distance calculation - dont use the thread
     if(kernelW < width)
     {
-        float pval = image_ptr[0][0][h][kernelW];
+        float l_dist;
         float new_dist = distance_ptr[0][0][h][kernelW];
         float cur_dist = 0.0;
         for (int w_i = 0; w_i < 3; w_i++)
@@ -87,7 +77,18 @@ __global__ void geodesic_updown_single_row_pass_kernel(
             const int kernelW_ind = kernelW + w_i - 1;
             if (kernelW_ind >= 0 && kernelW_ind < width)
             {
-                float l_dist = l1distance_cuda(pval, image_ptr[0][0][h + direction][kernelW_ind]);
+                l_dist = 0.0;
+                if (channel == 1)
+                {
+                    l_dist = l1distance_cuda(image_ptr[0][0][h][kernelW], image_ptr[0][0][h + direction][kernelW_ind]);
+                }
+                else
+                {
+                    for (int c_i = 0; c_i < channel; c_i++)
+                    {
+                        l_dist += l1distance_cuda(image_ptr[0][c_i][h][kernelW], image_ptr[0][c_i][h + direction][kernelW_ind]);
+                    }
+                }
                 cur_dist =  distance_ptr[0][0][h + direction][kernelW_ind] + l_eucl * local_dist2d[w_i] + l_grad * l_dist;
                 new_dist = std::min(new_dist, cur_dist);
             }
@@ -106,6 +107,7 @@ __global__ void geodesic_updown_single_row_pass_ptr_kernel(
     float l_eucl,
     const int direction,
     const int h,
+    const int channel,
     const int height,
     const int width
     )
@@ -115,16 +117,28 @@ __global__ void geodesic_updown_single_row_pass_ptr_kernel(
     // if outside, then skip distance calculation - dont use the thread
     if(kernelW < width)
     {
-        float pval = image_ptr[h * width + kernelW];
+        float l_dist;
         float new_dist = distance_ptr[h * width + kernelW];
         float cur_dist = 0.0;
 
         for (int w_i = 0; w_i < 3; w_i++)
         {
             const int kernelW_ind = kernelW + w_i - 1;
+            
             if (kernelW_ind >= 0 && kernelW_ind < width)
             {
-                float l_dist = l1distance_cuda(pval, image_ptr[(h + direction) * width + kernelW_ind]);
+                l_dist = 0.0;
+                if (channel == 1)
+                {
+                    l_dist = l1distance_cuda(image_ptr[h * width + kernelW], image_ptr[(h + direction) * width + kernelW_ind]);
+                }
+                else
+                {
+                    for (int c_i = 0; c_i < channel; c_i++)
+                    {
[…]
