### 2022-07-16 3824675
update demo2d to utilize cuda for multi-channel input
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
