### 2022-07-16 cda522c
refactoring for consistency
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
