+++ samples/demo2d.py
@@ -185,7 +185,7 @@ def demo_geodesic_distance2d(image):
 
 if __name__ == "__main__":
     # "gray" or "color"
-    example = "color"
+    example = "gray"
 
     if example == "gray":
         image = Image.open("data/img2d.png").convert("L")
