+++ FastGeodis/fastgeodis_cpu.cpp
@@ -58,27 +58,6 @@ float l1distance(const float &in1, const float &in2)
     return std::abs(in1 - in2);
 }
 
-
-float l1distance(const float *in1, const float *in2, int size)
-{
-    float ret_sum = 0.0;
-    for (int c_i = 0; c_i < size; c_i++)
-    {
-        ret_sum += abs(in1[c_i] - in2[c_i]);
-    }
-    return ret_sum;
-}
-
-float l1distance(const std::vector<float> &in1, const std::vector<float> &in2, int size)
-{
-    float ret_sum = 0.0;
-    for (int c_i = 0; c_i < size; c_i++)
-    {
-        ret_sum += abs(in1[c_i] - in2[c_i]);
-    }
-    return ret_sum;
-}
-
 // float l2distance(const float *in1, const float *in2, int size)
 // {
 //     float ret_sum = 0.0;
@@ -89,7 +68,6 @@ float l1distance(const std::vector<float> &in1, const std::vector<float> &in2, i
 //     return std::sqrt(ret_sum);
 // }
 
-
 void geodesic_updown_pass_cpu(const torch::Tensor &image, torch::Tensor &distance, const float &l_grad,  const float &l_eucl)
 {
     // batch, channel, height, width
@@ -99,7 +77,6 @@ void geodesic_updown_pass_cpu(const torch::Tensor &image, torch::Tensor &distanc
 
     auto image_ptr = image.accessor<float, 4>();
     auto distance_ptr = distance.accessor<float, 4>();
-    // constexpr float local_dist[] = {sqrt(float(2.)), float(1.), sqrt(float(2.))};
     const float local_dist[] = {sqrt(float(2.)), float(1.), sqrt(float(2.))};
 
     // top-down
@@ -111,27 +88,27 @@ void geodesic_updown_pass_cpu(const torch::Tensor &image, torch::Tensor &distanc
         #endif
         for (int w = 0; w < width; w++)
         {
+            float l_dist, cur_dist;
             float new_dist = distance_ptr[0][0][h][w];
             for (int w_i = 0; w_i < 3; w_i++)
             {
                 const int w_ind = w + w_i - 1;
                 if (w_ind < 0 || w_ind >= width)
                     continue;
-
-                float l_dist;
+                
+                l_dist = 0.0;
                 if (channel == 1)
                 {
                     l_dist = l1distance(image_ptr[0][0][h][w], image_ptr[0][0][h - 1][w_ind]);
                 }
                 else
                 {
-                    l_dist = 0;
                     for (int c_i = 0; c_i < channel; c_i++)
                     {
                         l_dist += l1distance(image_ptr[0][c_i][h][w], image_ptr[0][c_i][h - 1][w_ind]);
                     }
                 }
-                const float cur_dist = distance_ptr[0][0][h - 1][w_ind] + l_eucl * local_dist[w_i] + l_grad * l_dist;
+                cur_dist = distance_ptr[0][0][h - 1][w_ind] + l_eucl * local_dist[w_i] + l_grad * l_dist;
                 new_dist = std::min(new_dist, cur_dist);
             }
             distance_ptr[0][0][h][w] = new_dist;
@@ -147,28 +124,27 @@ void geodesic_updown_pass_cpu(const torch::Tensor &image, torch::Tensor &distanc
         #endif
         for (int w = 0; w < width; w++)
         {
+            float l_dist, cur_dist;
             float new_dist = distance_ptr[0][0][h][w];
             for (int w_i = 0; w_i < 3; w_i++)
             {
                 const int w_ind = w + w_i - 1;
                 if (w_ind < 0 || w_ind >= width)
                     continue;
-
-                float l_dist;
+                
+                l_dist = 0;
[…]
