### 2022-07-16 3575909
fix code formatting
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
