+++ FastGeodis/fastgeodis_cpu.cpp
@@ -35,24 +35,6 @@
 #include <omp.h>
 #endif
 
-// void print_shape(torch::Tensor data)
-// {
-//     auto num_dims = data.dim();
-//     std::cout << "Shape: (";
-//     for (int dim = 0; dim < num_dims; dim++)
-//     {
-//         std::cout << data.size(dim);
-//         if (dim != num_dims - 1)
-//         {
-//             std::cout << ", ";
-//         }
-//         else
-//         {
-//             std::cout << ")" << std::endl;
-//         }
-//     }
-// }
-
 float l1distance(const float &in1, const float &in2)
 {
     return std::abs(in1 - in2);
@@ -68,7 +50,12 @@ float l1distance(const float &in1, const float &in2)
 //     return std::sqrt(ret_sum);
 // }
 
-void geodesic_updown_pass_cpu(const torch::Tensor &image, torch::Tensor &distance, const float &l_grad,  const float &l_eucl)
+void geodesic_updown_pass_cpu(
+    const torch::Tensor &image, 
+    torch::Tensor &distance, 
+    const float &l_grad, 
+    const float &l_eucl
+    )
 {
     // batch, channel, height, width
     const int channel = image.size(1);
@@ -90,25 +77,36 @@ void geodesic_updown_pass_cpu(const torch::Tensor &image, torch::Tensor &distanc
         {
             float l_dist, cur_dist;
             float new_dist = distance_ptr[0][0][h][w];
+
             for (int w_i = 0; w_i < 3; w_i++)
             {
                 const int w_ind = w + w_i - 1;
+
                 if (w_ind < 0 || w_ind >= width)
                     continue;
-                
+
                 l_dist = 0.0;
                 if (channel == 1)
                 {
-                    l_dist = l1distance(image_ptr[0][0][h][w], image_ptr[0][0][h - 1][w_ind]);
+                    l_dist = l1distance(
+                        image_ptr[0][0][h][w], 
+                        image_ptr[0][0][h - 1][w_ind]
+                        );
                 }
                 else
                 {
                     for (int c_i = 0; c_i < channel; c_i++)
                     {
-                        l_dist += l1distance(image_ptr[0][c_i][h][w], image_ptr[0][c_i][h - 1][w_ind]);
+                        l_dist += l1distance(
+                            image_ptr[0][c_i][h][w], 
+                            image_ptr[0][c_i][h - 1][w_ind]
+                            );
                     }
                 }
-                cur_dist = distance_ptr[0][0][h - 1][w_ind] + l_eucl * local_dist[w_i] + l_grad * l_dist;
+                cur_dist = distance_ptr[0][0][h - 1][w_ind] + \
+                            l_eucl * local_dist[w_i] + \
+                            l_grad * l_dist;
+
                 new_dist = std::min(new_dist, cur_dist);
             }
             distance_ptr[0][0][h][w] = new_dist;
@@ -126,25 +124,36 @@ void geodesic_updown_pass_cpu(const torch::Tensor &image, torch::Tensor &distanc
         {
             float l_dist, cur_dist;
             float new_dist = distance_ptr[0][0][h][w];
+
             for (int w_i = 0; w_i < 3; w_i++)
             {
                 const int w_ind = w + w_i - 1;
+
                 if (w_ind < 0 || w_ind >= width)
                     continue;
-                
+
                 l_dist = 0;
                 if (channel == 1)
                 {
-                    l_dist = l1distance(image_ptr[0][0][h][w], image_ptr[0][0][h + 1][w_ind]);
+                    l_dist = l1distance(
+                        image_ptr[0][0][h][w], 
+                        image_ptr[0][0][h + 1][w_ind]
+                        );
                 }
                 else
                 {
[…]
