### 2022-07-17 5914d7e
add paper citation details
