### 2022-07-19 masadcv/FastGeodis#14 @jacobmerson
build error with GeodisTK (required for examples)
I'm working through the Joss review (https://github.com/openjournals/joss-reviews/issues/4532) and I am trying to run some of the samples scripts. These alll require `GeodisTK` which has build errors on Clang13.0.1.

I have opened an issue on the `GeodisTK` repo here: https://github.com/taigw/GeodisTK/issues/14
