### 2022-07-19 masadcv/FastGeodis#14 @masadcv
Hi @jacobmerson ,
Many thanks for reporting this, I will look into this..

GeodisTK is only used for baseline comparisons and hence (if needed) can be safely removed from samples
