### 2022-07-20 #4532 @yhtang
I am a bit concerned regarding whether this submission meets the **Substantial scholarly effort** criterion. My concerns are based on the following facts:

1. Lines of code. The main package folder `FastGeodis` contains 1090 lines of CUDA C++/Python code as of today, which seems to be just above the 1000 LOC rule-of-thumb in the review guideline. There are 307 and 490 additional lines of Pythons code in the `tests` and `samples` folder, respectively.
2. Amount of work. The package appears to be a set of OpenMP/CUDA kernels wrapped by a thin Python interface layer. Combining with the lines of code info above, I am not convinced yet that this represents 3 months of work for a full-time contributor.

I wonder what our editor and other reviewers think regarding this?
