### 2022-07-20 #4532 @masadcv
> @masadcv Could you please also point me to the specific example where the 15x speedup over CPU and 60x speedup over GPU claim in the statement of need can be reproduced? What was the specification of the machine/platform where the numbers are obtained?

Hi @yhtang ,
Yes of course.
The script used for getting these figures is here: [samples/test_speed_benchmark.py](https://github.com/masadcv/FastGeodis/blob/master/samples/test_speed_benchmark.py). When you run it, it will output two figures found here: [figures/experiment_2d.png](https://github.com/masadcv/FastGeodis/blob/master/figures/experiment_2d.png) and here [figures/experiment_3d.png](https://github.com/masadcv/FastGeodis/blob/master/figures/experiment_3d.png)

We ran this on a machine with:
- GPU:  Nvidia GeForce Titan X with 12 GB
- 6-Core Intel Xeon E5-1650 v3

And got the results
![experiment_2d](https://user-images.githubusercontent.com/3410225/180062697-8326f9c2-1061-42f7-bf19-da7cfc917889.png)
![experiment_3d](https://user-images.githubusercontent.com/3410225/180062704-4a4a0270-0302-46a7-b560-a64b9d539795.png)

For both 2d and 3d cases we report speedup of upto 15x and 60x by looking at the maximum (upto) improvement when comparing against existing library (GeodisTK). This occurs at the biggest spatial dimension in the above figures.

I hope this helps. Please do not hesitate to reach out if anything is unclear in above.
