### 2022-07-20 #4532 @masadcv
> Regarding the documentation and the paper, I think it will really help if a definition can be given for geodesic transform, in both layman's terms and math formulas. Otherwise, I (and potential future users) can only guess what it is but have to infer what exactly is done by looking at the code.

Many thanks for pointing this out. The main equation is found here: https://github.com/masadcv/FastGeodis/pull/2#issue-962801700
We will add this and layman description of distance transforms (including geodesic and euclidean distance transforms) to documentation and to the paper as well.
