### 2022-07-21 #4532 @jacobmerson
> I am a bit concerned regarding whether this submission meets the **Substantial scholarly effort** criterion. My concerns are based on the following facts:
>
> 1. Lines of code. The main package folder `FastGeodis` contains 1090 lines of CUDA C++/Python code as of today, which seems to be just above the 1000 LOC rule-of-thumb in the review guideline. There are 307 and 490 additional lines of Pythons code in the `tests` and `samples` folder, respectively.
> 2. Amount of work. The package appears to be a set of OpenMP/CUDA kernels wrapped by a thin Python interface layer. Combining with the lines of code info above, I am not convinced yet that this represents 3 months of work for a full-time contributor.
>
> I wonder what our editor and other reviewers think regarding this?

I think this library may be on the edge in terms of "scholarly effort" w.r.t. the code. In general writing performant cuda kernels can be quite time consuming. I looked through the commit history and it looks like most of the development was done in a few short bursts a few days at a time. Now, this can be misleading because often work gets done that sits around and doesn't get committed back until there are deadlines on the table.

I think the authors have done done a good job of providing lots of examples and test cases.

I agree that it might be helpful to have an editor weigh in here.
