### 2022-07-21 masadcv/FastGeodis#14 @jacobmerson
@masadcv I submitted a pull request to GeodisTK with the fixes for those errors. https://github.com/taigw/GeodisTK/pull/15
