### 2022-07-21 #4532 @jacobmerson
Paper comments: @masadcv

Summary:
You mention that Geodesic distance transforms are important for image segmentation. If this is for a non-expert audience you should mention what is segmentation and why it is important.

You need to specify what is a Geodesic distance transform and how it differs from the Euclidean distance transform.

I am in agreement with @yhtang if you want to make the speedup claim in the paper you need to provide more information (in the paper). At a minimum you need to provide the hardware information you provided in your comment here.

On the CPU are you comparing a single thread to a full GPU? Or something different.

You need to mention if your performance is relevant for 2D or 3D.

Additionally, in the summary section I think you need to give at least a hint as to why you are seeing better performance in both your CPU and GPU implementations. You say “efficient use of hardware” does this just correspond to use multithreading, are you doing something to minimize cache misses, etc.

Based on the Joss guidelines, it’s not clear that the methodology section is needed in the paper.

However, it would be very useful to include a more fleshed out version of this in the documentation.

You need to be more specific about what the algorithm is and how you implement it in code. You mention that there are two passes needed in 2D and 6 passes needed in 3D. Are those passes computed in parallel? How do you break up you algorithm into independent chunks? What quantity are you computing (what’s the equation). What considerations need to be made in the CPU vs GPU implementation (i.e., do you use the same algorithm and is the parallelization scheme identical)?
