### 2022-07-21 #4532 @masadcv
Many thanks @jacobmerson , your feedback is much appreciated. I am working on improving the paper based on your and @yhtang suggestions. Below is my response to some of your questions, which I will include in paper.

> On the CPU are you comparing a single thread to a full GPU? Or something different.
>
> You need to mention if your performance is relevant for 2D or 3D.
>
The main comparison is with existing open-source library that perform similar operation (GeodisTK). We report performance numbers as compared to this. GeodisTK implements with a single thread where the algorithm is not directly parallelisable, while our method can utilise multiple thread on CPU using OpenMP.  If one has GPU, then the performance gain becomes much greater (as reported in the paper). I will explicitly mention what performance improvements can be gained on each and with 2D and 3D data and also include the hardware information for these tests.

> Additionally, in the summary section I think you need to give at least a hint as to why you are seeing better performance in both your CPU and GPU implementations. You say “efficient use of hardware” does this just correspond to use multithreading, are you doing something to minimize cache misses, etc.

You are right, this performance improvement is coming from parallelisable algorithm that runs in multiple threads/cuda cores. This was not the case in previous open source solutions.

> Based on the Joss guidelines, it’s not clear that the methodology section is needed in the paper.
>
> However, it would be very useful to include a more fleshed out version of this in the documentation.
>
> You need to be more specific about what the algorithm is and how you implement it in code. You mention that there are two passes needed in 2D and 6 passes needed in 3D. Are those passes computed in parallel? How do you break up you algorithm into independent chunks? What quantity are you computing (what’s the equation).

In each pass, the algorithm can do work in parallel on a row (for 2D) and a plane (for 3D). That is the only place where we do parallelisation. We run multiple passes of this in each direction to compute distance transform. These passes itself are run in sequence, as they require output from the previous pass to proceed. I will include the equation in paper. I will make this clear in paper.

> What considerations need to be made in the CPU vs GPU implementation (i.e., do you use the same algorithm and is the parallelization scheme identical)?

Looking at the overall algorithm, both CPU and GPU are parallising the for loop over each element in row/plane. For CPU, the load for a row/plane is split into multiple threads. For GPU, this idea is same in principle, but involves splitting workload into multiple blocks and each block has multiple threads. GPU particularly helps in 3D where we can split the task over the plane into multiple blocks.
