### 2022-07-21 #4532 @vissarion
Thanks @yhtang for your raising this question about scope and thanks @jacobmerson for your feedback. I personally found the scholarly effort on the edge especially on the length of the core source code. Thus following the guidelines
```
The decision on scholarly effort is ultimately one made by JOSS editors.
Reviewers are asked to flag submissions of questionable scope during the review process so that
the editor can bring this to the attention of the JOSS editorial team.
```
@masadcv I am bringing this issue to the JOSS editorial team to seek the input of a few of our editors. This process will likely take a week or so.
