### 2022-07-21 e037a41
remove redundant macos 10 tests
