+++ .github/workflows/build-sdist-publish.yml
@@ -17,7 +17,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [windows-2019, macos-10.15, macos-12, ubuntu-18.04]
+        os: [windows-2019, macos-12, ubuntu-18.04]
         python-version: [3.6, 3.7, 3.8, 3.9]
     steps:
     - uses: actions/checkout@v1
