### 2022-07-21 0954b40
remove redundant macos 10 tests
