### 2022-07-21 #4532 @masadcv
Many thanks @vissarion for suggesting scope check with editorial team. Below is our comment for consideration of the editorial review:

As rightly picked by @jacobmerson, CUDA code requires quite a lot of effort and time to apply to any given problem. One potential drawback of getting things right is that it may result in less lines of code.

The main contribution here is the impact of how much improvement it brings as compared to existing open source libraries. As reported above and in the paper, our tests reveal this to be significant impact - resulting in application of Geodesic distance in real-time applications, which was not possible before. In fact, this package is already used by a number of researchers (including myself) to enable real-time applications for medical imaging. We think this will be relevant for the wider community as well, and for this reason we have released the code and made a submission as a JOSS paper.
