### 2022-07-21 55185fe
improve installation and unittest instructions
