### 2022-07-22 #4532 @yhtang
The CUDA kernel appears to be straightforward stencil code without much in-depth optimization. They are memory-bound because each CUDA thread repeatedly loads the pixel data for distance computation without sharing and reusing. The arithmetic intensity (amount of computation per unit of data) is very low. Patterns such as shared memory tiling, register blocking, and warp shuffle are commonly used to optimize this type of workload, but do not appear to be implemented here. Thus, the observed speedup is likely determined by the difference in memory bandwidth, rather than FLOPS, between the CPU and the GPU.

More interestingly, I just ran a simplified version of the [Simple 2D Geodesic and Euclidean Distance](https://colab.research.google.com/github/masadcv/FastGeodis/blob/master/samples/simpledemo2d.ipynb) example as follow and found that the `geodesic_updown_single_row_pass_kernel` kernel gets launched by **1336 times**, each time with a 1x1x1 grid and 256x1x1 block. This could be easily converted to a single launch with a 1336x1x1 grid and 256x1x1 block size to avoid the huge kernel launch overhead currently invoked. This is further evidence that the software in its current form lacks a certain amount of maturity.

```
import os
import wget
import numpy as np
from PIL import Image
import matplotlib.pyplot as plt
import torch
import FastGeodis

# read images from local path or download from github repo (for colab)
image_path = "img2d.png"
url = "https://raw.githubusercontent.com/masadcv/FastGeodis/master/data/"
wget.download(os.path.join(url, image_path), image_path) if not os.path.exists(image_path) else None

device = "cuda"
image = np.asarray(Image.open(image_path), np.float32)

image_pt = torch.from_numpy(image).unsqueeze_(0).unsqueeze_(0)
image_pt = image_pt.to(device)
mask_pt = torch.ones_like(image_pt)
mask_pt[..., 100, 100] = 0

v = 1e10
iterations = 2

lamb = 1.0 # <-- Geodesic distance transform
geodesic_dist = FastGeodis.generalised_geodesic2d(
    image_pt, mask_pt, v, lamb, iterations
)
geodesic_dist = np.squeeze(geodesic_dist.cpu().numpy())

print(geodesic_dist)
```
