+++ FastGeodis/geodis_toivanen.cpp
@@ -1,0 +1,371 @@
+// BSD 3-Clause License
+
+// Copyright (c) 2021, Muhammad Asad (masadcv@gmail.com)
+// All rights reserved.
+
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+
+// 1. Redistributions of source code must retain the above copyright notice, this
+//    list of conditions and the following disclaimer.
+
+// 2. Redistributions in binary form must reproduce the above copyright notice,
+//    this list of conditions and the following disclaimer in the documentation
+//    and/or other materials provided with the distribution.
+
+// 3. Neither the name of the copyright holder nor the names of its
+//    contributors may be used to endorse or promote products derived from
+//    this software without specific prior written permission.
+
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+// DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+// OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+#include <torch/extension.h>
+#include <vector>
+// #include <iostream>
+
+float l1distance_toivanen(const float &in1, const float &in2)
+{
+    return std::abs(in1 - in2);
+}
+
+// float l2distance(const float *in1, const float *in2, int size)
+// {
+//     float ret_sum = 0.0;
+//     for (int c_i = 0; c_i < size; c_i++)
+//     {
+//         ret_sum += (in1[c_i] - in2[c_i]) * (in1[c_i] - in2[c_i]);
+//     }
+//     return std::sqrt(ret_sum);
+// }
+
+void geodesic2d_forback_toivanen_cpu(
+    const torch::Tensor &image, 
+    torch::Tensor &distance, 
+    const float &l_grad, 
+    const float &l_eucl
+    )
+{
+    // batch, channel, height, width
+    const int channel = image.size(1);
+    const int height = image.size(2);
+    const int width = image.size(3);
+
+    auto image_ptr = image.accessor<float, 4>();
+    auto distance_ptr = distance.accessor<float, 4>();
+
+    // forward
+    const int dh_f[4] = {-1, -1, -1, 0};
+    const int dw_f[4] = {-1, 0, 1, -1};
+
+    const float local_dist_f[] = {sqrt(float(2.)), float(1.), sqrt(float(2.)), float(1.)};
+
+    for (int h = 0; h < height; h++)
+    {
+        for (int w = 0; w < width; w++)
+        {
+            float l_dist, cur_dist;
+            float new_dist = distance_ptr[0][0][h][w];
+
+            for (int ind = 0; ind < 4; ind++)
[…]
