+++ tests/test_fastgeodis.py
@@ -78,6 +78,15 @@ def fastgeodis_generalised_geodesic_distance_3d(
 ):
     return FastGeodis.generalised_geodesic3d(image, softmask, spacing, v, lamb, iter)
 
+def toivanen_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
+    return FastGeodis.generalised_geodesic2d_toivanen(image, softmask, v, lamb, iter)
+
+
+def toivanen_generalised_geodesic_distance_3d(
+    image, softmask, v, lamb, iter, spacing
+):
+    return FastGeodis.generalised_geodesic3d_toivanen(image, softmask, spacing, v, lamb, iter)
+
 
 def fastgeodis_GSF_2d(image, softmask, theta, v, lamb, iter):
     return FastGeodis.GSF2d(image, softmask, theta, v, lamb, iter)
@@ -101,6 +110,14 @@ def get_fastgeodis_func(num_dims, spacing=[1.0, 1.0, 1.0]):
     else:
         raise ValueError("Unsupported num_dims received: {}".format(num_dims))
 
+def get_geodis_toivanen_func(num_dims, spacing=[1.0, 1.0, 1.0]):
+    if num_dims == 2:
+        return toivanen_generalised_geodesic_distance_2d
+    elif num_dims == 3:
+        return partial(toivanen_generalised_geodesic_distance_3d, spacing=spacing)
+    else:
+        raise ValueError("Unsupported num_dims received: {}".format(num_dims))
+
 
 def get_GSF_func(num_dims, spacing=[1.0, 1.0, 1.0]):
     if num_dims == 2:
@@ -111,11 +128,17 @@ def get_GSF_func(num_dims, spacing=[1.0, 1.0, 1.0]):
         raise ValueError("Unsupported num_dims received: {}".format(num_dims))
 
 
-DEVICES_TO_RUN = ["cpu", "cuda"]
-CONF_2D = [(dev, 2, bas) for dev in DEVICES_TO_RUN for bas in [16, 32, 64]]
-CONF_3D = [(dev, 3, bas) for dev in DEVICES_TO_RUN for bas in [16, 32, 64]]
+CONF_2D_CPU = [("cpu", 2, bas) for bas in [16, 32, 64]]
+CONF_2D_CUDA = [("cuda", 2, bas) for bas in [16, 32, 64]]
+CONF_2D = CONF_2D_CPU + CONF_2D_CUDA
+
+CONF_3D_CPU = [("cpu", 3, bas) for bas in [16, 32, 64]]
+CONF_3D_CUDA = [("cuda", 3, bas) for bas in [16, 32, 64]]
+CONF_3D = CONF_3D_CPU + CONF_2D_CUDA
+
 CONF_ALL = CONF_2D + CONF_3D
 
+CONF_ALL_CPU = CONF_2D_CPU + CONF_3D_CPU
 
 class TestFastGeodis(unittest.TestCase):
     @parameterized.expand(CONF_ALL)
@@ -285,7 +308,7 @@ class TestFastGeodis(unittest.TestCase):
         self.assertTrue(check)
 
     @parameterized.expand(CONF_3D)
-    @run_cuda_if_available
+    # @run_cuda_if_available
     def test_ill_spacing(self, device, num_dims, base_dim):
         print(device)
         print(num_dims)
@@ -326,188 +349,386 @@ class TestFastGeodis(unittest.TestCase):
                 mask = torch.rand(mask_shape_mod, dtype=torch.float32).to(device)
                 geodesic_dist = geodis_func(image, mask, 1e10, 1.0, 2)
 
-class TestFastGeodisSigned(unittest.TestCase):
-    @parameterized.expand(CONF_ALL)
-    @run_cuda_if_available
[…]

[…]
