### 2022-07-23 #4532 @masadcv
Hi @yhtang

Many thanks for your detailed analysis of our 2D case. Really appreciate your time and effort looking into this and reporting the findings.

I believe the main algorithm for our approach may be misunderstood here. Although we have a full image data, we can only process one row/plane at a time and can only parallelize workload over this row/plane. This is unlike other image processing kernels than dont have such data dependency and hence can operate on all points within the image.

Let me try to explain the algorithm first, and then I will explain why we need 1336 invocations in the example you analyzed. So the 2d algorithm works by computing distance propagation in one row at a time. This is a hard constraint because compute for each row is dependent on computed distances in previous rows from previous invocation. Consider an example of an image with 4 x 6 dimension. Then a successful full iteration of our method would involve going through the rows one by one as follows:

![output](https://user-images.githubusercontent.com/3410225/180580833-fc72cfbe-834e-49e3-8fbf-339bb991df6e.gif)

Alternate video link: https://www.youtube.com/watch?v=tj7yS3AQbJI

As you can see, this involves top-down and left-right passes. How we implement left-right is by reusing top-down code and transposing the data instead. What is important in the above is in each invocation, **only the pixels highlighted in green color can be computed** (as they have the data available from previous row). It is this row that we split into multiple threads (what you see as 256 blocks in profiler). If we calculate the total number of invocations for this simple example, then they are:

![Slide28](https://user-images.githubusercontent.com/3410225/180581014-71b18253-944b-4c00-8b6d-b044b3c61e43.PNG)

Going back to the simple example 2d case you ran through profiler. In this, the image size is 128 x 208 and the whole algorithm is repeated twice (iterations=2). If we calculate the number of rows (or invocations) that can be worked out fully at each row pass, then this comes out to be:
![Slide29 (copy)](https://user-images.githubusercontent.com/3410225/180582156-99fdb3d3-0c04-4464-9fea-1e96be0fd1b3.PNG)

Given that we have boundary conditions that skip passes that don't have data, when you look at the profiler you see 1336 invocations.

As you rightly pointed out, there may be alternate ways of implementing this, for example a kernel that internally goes over all invocation steps above. However this would lead to additional issues (e.g. race condition, synchronization of threads, transpose of data etc) which may need to be addressed with additional overheads.

Going beyond just 2D examples, if you look at our 3D code, we have cuda kernels that operate on the same principle (we can process one plane at a time). In that case, since we have more data, we can utilize more compute on GPU and process a plane in parallel, however we still have the data dependency constraints that prevent us from processing all planes together.

The above described algorithm is not something we came up with. It is implemented from ref [1] as noted in our paper and documentation. Here is the original slides from that work that we used for guidance:

![Screenshot from 2022-07-23 01-03-48](https://user-images.githubusercontent.com/3410225/180582106-298d32cc-d627-4637-8c42-ff214484fae9.png)
![Screenshot from 2022-07-23 01-04-04](https://user-images.githubusercontent.com/3410225/180582112-b0694d1c-05bb-4553-9c49-099f62ec28f8.png)

[1] Criminisi, Antonio, Toby Sharp, and Khan Siddiqui. "Interactive Geodesic Segmentation of n-Dimensional Medical Images on the Graphics Processor." Radiological Society of North America (RSNA), 2009. [[pdf](https://www.microsoft.com/en-us/research/wp-content/uploads/2016/02/Criminisi_RSNA09_extendedAbstract1.pdf), [ppt](https://www.microsoft.com/en-us/research/wp-content/uploads/2016/02/RSNA2009_Segmentation_Criminisi.ppt), [url](https://www.microsoft.com/en-us/research/publication/interactive-geodesic-segmentation-of-n-dimensional-medical-images-on-the-graphics-processor/)]
