### 2022-07-23 masadcv/FastGeodis#14 @masadcv
@jacobmerson many thanks!

I have opened a PR to add a comparison Geodesic distance method that we can use for comparisons in samples and hence become independent of external GeodisTK package.

This is still a WIP, but I will try to get this in during the weekend..
