+++ tests/test_fastgeodis.py
@@ -79,6 +79,27 @@ def fastgeodis_generalised_geodesic_distance_3d(
     return FastGeodis.generalised_geodesic3d(image, softmask, spacing, v, lamb, iter)
 
 
+def fastgeodis_signed_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
+    return FastGeodis.signed_generalised_geodesic2d(image, softmask, v, lamb, iter)
+
+
+def fastgeodis_signed_generalised_geodesic_distance_3d(
+    image, softmask, v, lamb, iter, spacing
+):
+    return FastGeodis.signed_generalised_geodesic3d(image, softmask, spacing, v, lamb, iter)
+
+
+def toivenan_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
+    return FastGeodis.generalised_geodesic2d_toivanen(image, softmask, v, lamb, iter)
+
+
+def toivenan_generalised_geodesic_distance_3d(
+    image, softmask, v, lamb, iter, spacing
+):
+    return FastGeodis.generalised_geodesic3d_toivanen(image, softmask, spacing, v, lamb, iter)
+
+
+
 def fastgeodis_GSF_2d(image, softmask, theta, v, lamb, iter):
     return FastGeodis.GSF2d(image, softmask, theta, v, lamb, iter)
 
@@ -101,6 +122,22 @@ def get_fastgeodis_func(num_dims, spacing=[1.0, 1.0, 1.0]):
     else:
         raise ValueError("Unsupported num_dims received: {}".format(num_dims))
 
+def get_signed_fastgeodis_func(num_dims, spacing=[1.0, 1.0, 1.0]):
+    if num_dims == 2:
+        return fastgeodis_signed_generalised_geodesic_distance_2d
+    elif num_dims == 3:
+        return partial(fastgeodis_signed_generalised_geodesic_distance_3d, spacing=spacing)
+    else:
+        raise ValueError("Unsupported num_dims received: {}".format(num_dims))
+
+def get_toivenan_func(num_dims, spacing=[1.0, 1.0, 1.0]):
+    if num_dims == 2:
+        return toivenan_generalised_geodesic_distance_2d
+    elif num_dims == 3:
+        return partial(toivenan_generalised_geodesic_distance_3d, spacing=spacing)
+    else:
+        raise ValueError("Unsupported num_dims received: {}".format(num_dims))
+
 
 def get_GSF_func(num_dims, spacing=[1.0, 1.0, 1.0]):
     if num_dims == 2:
@@ -112,10 +149,16 @@ def get_GSF_func(num_dims, spacing=[1.0, 1.0, 1.0]):
 
 
 DEVICES_TO_RUN = ["cpu", "cuda"]
-CONF_2D = [(dev, 2, bas) for dev in DEVICES_TO_RUN for bas in [16, 32, 64]]
-CONF_3D = [(dev, 3, bas) for dev in DEVICES_TO_RUN for bas in [16, 32, 64]]
+CONF_2D_CPU = [("cpu", 2, bas) for bas in [16, 32, 64]]
+CONF_2D_CUDA = [("cuda", 2, bas) for bas in [16, 32, 64]]
+CONF_2D = CONF_2D_CPU + CONF_2D_CUDA
+
+CONF_3D_CPU = [("cpu", 3, bas) for bas in [16, 32, 64]]
+CONF_3D_CUDA = [("cuda", 3, bas) for bas in [16, 32, 64]]
+CONF_3D = CONF_3D_CPU + CONF_3D_CUDA
 CONF_ALL = CONF_2D + CONF_3D
 
+CONF_ALL_CPU = CONF_2D_CPU + CONF_3D_CPU
 
[…]
