### 2022-07-23 d0d2602
add toivanens additional functions and complete unittests
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
