+++ FastGeodis/fastgeodis.cpp
@@ -218,16 +218,64 @@ torch::Tensor GSF3d(torch::Tensor &image, const torch::Tensor &mask, const float
     return Dd_Md + De_Me;
 }
 
+torch::Tensor getDs2d_toivanen(torch::Tensor &image, const torch::Tensor &mask, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
+{
+    torch::Tensor D_M = generalised_geodesic2d_toivanen(image, mask, v, l_grad, l_eucl, iterations);
+    torch::Tensor D_Mb = generalised_geodesic2d_toivanen(image, 1 - mask, v, l_grad, l_eucl, iterations);
+
+    return D_M - D_Mb;
+}
+
+torch::Tensor GSF2d_toivanen(torch::Tensor &image, const torch::Tensor &mask, const float &theta, const float &v, const float &lambda, const int &iterations)
+{
+    torch::Tensor Ds_M = getDs2d_toivanen(image, mask, v, lambda, 1 - lambda, iterations);
+
+    torch::Tensor Md = (Ds_M > theta).type_as(Ds_M);
+    torch::Tensor Me = (Ds_M > -theta).type_as(Ds_M);
+
+    torch::Tensor Dd_Md = -getDs2d_toivanen(image, 1 - Md, v, lambda, 1 - lambda, iterations);
+    torch::Tensor De_Me = getDs2d_toivanen(image, Me, v, lambda, 1 - lambda, iterations);
+
+    return Dd_Md + De_Me;
+}
+
+torch::Tensor getDs3d_toivanen(torch::Tensor &image, const torch::Tensor &mask, const std::vector<float> &spacing, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
+{
+    torch::Tensor D_M = generalised_geodesic3d_toivanen(image, mask, spacing, v, l_grad, l_eucl, iterations);
+    torch::Tensor D_Mb = generalised_geodesic3d_toivanen(image, 1 - mask, spacing, v, l_grad, l_eucl, iterations);
+
+    return D_M - D_Mb;
+}
+
+torch::Tensor GSF3d_toivanen(torch::Tensor &image, const torch::Tensor &mask, const float &theta, const std::vector<float> &spacing, const float &v, const float &lambda, const int &iterations)
+{
+    torch::Tensor Ds_M = getDs3d_toivanen(image, mask, spacing, v, lambda, 1 - lambda, iterations);
+
+    torch::Tensor Md = (Ds_M > theta).type_as(Ds_M);
+    torch::Tensor Me = (Ds_M > -theta).type_as(Ds_M);
+
+    torch::Tensor Dd_Md = -getDs3d_toivanen(image, 1 - Md, spacing, v, lambda, 1 - lambda, iterations);
+    torch::Tensor De_Me = getDs3d_toivanen(image, Me, spacing, v, lambda, 1 - lambda, iterations);
+
+    return Dd_Md + De_Me;
+}
 
 PYBIND11_MODULE(TORCH_EXTENSION_NAME, m)
 {
     m.def("generalised_geodesic2d", &generalised_geodesic2d, "Generalised Geodesic distance 2d");
     m.def("GSF2d", &GSF2d, "Geodesic Symmetric Filtering 2d");
     m.def("signed_generalised_geodesic2d", &getDs2d, "Signed Generalised Geodesic distance 2d");
-    m.def("generalised_geodesic2d_toivanen", &generalised_geodesic2d_toivanen, "Generalised Geodesic distance 2d using Toivanen's method");
 
+    m.def("generalised_geodesic2d_toivanen", &generalised_geodesic2d_toivanen, "Generalised Geodesic distance 2d using Toivanen's method");
[…]
