+++ FastGeodis/__init__.py
@@ -235,6 +235,78 @@ def generalised_geodesic3d_toivanen(
         image, softmask, spacing, v, lamb, 1 - lamb, iter
     )
 
+def signed_generalised_geodesic2d_toivanen(
+    image: torch.Tensor, 
+    softmask: torch.Tensor, 
+    v: float, 
+    lamb: float, 
+    iter: int = 2
+):
+    r"""Computes Signed Generalised Geodesic Distance using Toivanen's raster scanning method from:
+
+    Toivanen, Pekka J. 
+    "New geodosic distance transforms for gray-scale images." 
+    Pattern Recognition Letters 17.5 (1996): 437-450.
+
+    For more details on generalised geodesic distance, check the following reference:
+
+    Criminisi, Antonio, Toby Sharp, and Andrew Blake.
+    "Geos: Geodesic image segmentation."
+    European Conference on Computer Vision, Berlin, Heidelberg, 2008.
+
+    The function expects input as torch.Tensor, which can be run on CPU or GPU depending on Tensor's device location
+
+    Args:
+        image: input image, can be grayscale or multiple channels.
+        softmask: softmask in range [0, 1] with seed information.
+        v: weighting factor for establishing relationship between unary and spatial distances.
+        lamb: weighting factor between 0.0 and 1.0. 0.0 returns euclidean distance, whereas 1.0 returns geodesic distance
+        iter: number of passes of the iterative distance transform method
+
+    Returns:
+        torch.Tensor with distance transform
+    """
+    return FastGeodisCpp.signed_generalised_geodesic2d_toivanen(
+        image, softmask, v, lamb, 1 - lamb, iter
+    )
+
+
+def signed_generalised_geodesic3d_toivanen(
+    image: torch.Tensor,
+    softmask: torch.Tensor,
+    spacing: List,
+    v: float,
+    lamb: float,
+    iter: int = 4,
+):
+    r"""Computes Signed Generalised Geodesic Distance using Toivanen's raster scanning method from:
+
+    Toivanen, Pekka J. 
+    "New geodosic distance transforms for gray-scale images." 
+    Pattern Recognition Letters 17.5 (1996): 437-450.
+
+    For more details on generalised geodesic distance, check the following reference:
+
+    Criminisi, Antonio, Toby Sharp, and Andrew Blake.
+    "Geos: Geodesic image segmentation."
+    European Conference on Computer Vision, Berlin, Heidelberg, 2008.
+
+    The function expects input as torch.Tensor, which can be run on CPU or GPU depending on Tensor's device location
+
+    Args:
+        image: input image, can be grayscale or multiple channels.
+        softmask: softmask in range [0, 1] with seed information.
+        spacing: spacing for 3D data
+        v: weighting factor for establishing relationship between unary and spatial distances.
+        lamb: weighting factor between 0.0 and 1.0. 0.0 returns euclidean distance, whereas 1.0 returns geodesic distance
+        iter: number of passes of the iterative distance transform method
+
+    Returns:
+        torch.Tensor with distance transform
+    """
+    return FastGeodisCpp.signed_generalised_geodesic3d_toivanen(
+        image, softmask, spacing, v, lamb, 1 - lamb, iter
+    )
 
 def GSF2d(
     image: torch.Tensor,
@@ -296,3 +368,74 @@ def GSF3d(
         torch.Tensor with distance transform
     """
     return FastGeodisCpp.GSF3d(image, softmask, theta, spacing, v, lamb, iter)
+
+def GSF2d_toivanen(
+    image: torch.Tensor,
+    softmask: torch.Tensor,
+    theta: float,
+    v: float,
+    lamb: float,
+    iter: int,
+):
[…]

[…]
