+++ FastGeodis/geodis_toivanen.cpp
@@ -37,22 +37,11 @@ float l1distance_toivanen(const float &in1, const float &in2)
     return std::abs(in1 - in2);
 }
 
-// float l2distance(const float *in1, const float *in2, int size)
-// {
-//     float ret_sum = 0.0;
-//     for (int c_i = 0; c_i < size; c_i++)
-//     {
-//         ret_sum += (in1[c_i] - in2[c_i]) * (in1[c_i] - in2[c_i]);
-//     }
-//     return std::sqrt(ret_sum);
-// }
-
 void geodesic2d_forback_toivanen_cpu(
-    const torch::Tensor &image, 
-    torch::Tensor &distance, 
-    const float &l_grad, 
-    const float &l_eucl
-    )
+    const torch::Tensor &image,
+    torch::Tensor &distance,
+    const float &l_grad,
+    const float &l_eucl)
 {
     // batch, channel, height, width
     const int channel = image.size(1);
@@ -87,23 +76,21 @@ void geodesic2d_forback_toivanen_cpu(
                 if (channel == 1)
                 {
                     l_dist = l1distance_toivanen(
-                        image_ptr[0][0][h][w], 
-                        image_ptr[0][0][h_ind][w_ind]
-                        );
+                        image_ptr[0][0][h][w],
+                        image_ptr[0][0][h_ind][w_ind]);
                 }
                 else
                 {
                     for (int c_i = 0; c_i < channel; c_i++)
                     {
                         l_dist += l1distance_toivanen(
-                            image_ptr[0][c_i][h][w], 
-                            image_ptr[0][c_i][h_ind][w_ind]
-                            );
+                            image_ptr[0][c_i][h][w],
+                            image_ptr[0][c_i][h_ind][w_ind]);
                     }
                 }
-                cur_dist = distance_ptr[0][0][h_ind][w_ind] + \
-                            l_eucl * local_dist_f[ind] + \
-                            l_grad * l_dist;
+                cur_dist = distance_ptr[0][0][h_ind][w_ind] +
+                           l_eucl * local_dist_f[ind] +
+                           l_grad * l_dist;
 
                 new_dist = std::min(new_dist, cur_dist);
             }
@@ -119,7 +106,7 @@ void geodesic2d_forback_toivanen_cpu(
 
     for (int h = height - 1; h >= 0; h--)
     {
-        for (int w = width -1 ; w >= 0; w--)
+        for (int w = width - 1; w >= 0; w--)
         {
             float l_dist, cur_dist;
             float new_dist = distance_ptr[0][0][h][w];
@@ -136,24 +123,22 @@ void geodesic2d_forback_toivanen_cpu(
                 if (channel == 1)
                 {
                     l_dist = l1distance_toivanen(
-                        image_ptr[0][0][h][w], 
-                        image_ptr[0][0][h_ind][w_ind]
-                        );
+                        image_ptr[0][0][h][w],
+                        image_ptr[0][0][h_ind][w_ind]);
                 }
                 else
                 {
                     for (int c_i = 0; c_i < channel; c_i++)
                     {
                         l_dist += l1distance_toivanen(
-                            image_ptr[0][c_i][h][w], 
-                            image_ptr[0][c_i][h_ind][w_ind]
-                            );
+                            image_ptr[0][c_i][h][w],
+                            image_ptr[0][c_i][h_ind][w_ind]);
                     }
                 }
[…]
