### 2022-07-23 c7ff9e4
update unittest code formatting
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
