+++ tests/test_fastgeodis.py
@@ -86,28 +86,33 @@ def fastgeodis_signed_generalised_geodesic_distance_2d(image, softmask, v, lamb,
 def fastgeodis_signed_generalised_geodesic_distance_3d(
     image, softmask, v, lamb, iter, spacing
 ):
-    return FastGeodis.signed_generalised_geodesic3d(image, softmask, spacing, v, lamb, iter)
+    return FastGeodis.signed_generalised_geodesic3d(
+        image, softmask, spacing, v, lamb, iter
+    )
+
 
 def toivanen_signed_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
-    return FastGeodis.signed_generalised_geodesic2d_toivanen(image, softmask, v, lamb, iter)
+    return FastGeodis.signed_generalised_geodesic2d_toivanen(
+        image, softmask, v, lamb, iter
+    )
 
 
 def toivanen_signed_generalised_geodesic_distance_3d(
     image, softmask, v, lamb, iter, spacing
 ):
-    return FastGeodis.signed_generalised_geodesic3d_toivanen(image, softmask, spacing, v, lamb, iter)
-
+    return FastGeodis.signed_generalised_geodesic3d_toivanen(
+        image, softmask, spacing, v, lamb, iter
+    )
 
 
 def toivanen_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
     return FastGeodis.generalised_geodesic2d_toivanen(image, softmask, v, lamb, iter)
 
 
-def toivanen_generalised_geodesic_distance_3d(
-    image, softmask, v, lamb, iter, spacing
-):
-    return FastGeodis.generalised_geodesic3d_toivanen(image, softmask, spacing, v, lamb, iter)
-
+def toivanen_generalised_geodesic_distance_3d(image, softmask, v, lamb, iter, spacing):
+    return FastGeodis.generalised_geodesic3d_toivanen(
+        image, softmask, spacing, v, lamb, iter
+    )
 
 
 def fastgeodis_GSF_2d(image, softmask, theta, v, lamb, iter):
@@ -116,6 +121,7 @@ def fastgeodis_GSF_2d(image, softmask, theta, v, lamb, iter):
 
 def fastgeodis_GSF_3d(image, softmask, theta, v, lamb, iter, spacing):
     return FastGeodis.GSF3d(image, softmask, theta, spacing, v, lamb, iter)
+
 
 def toivanen_GSF_2d(image, softmask, theta, v, lamb, iter):
     return FastGeodis.GSF2d_toivanen(image, softmask, theta, v, lamb, iter)
@@ -139,13 +145,17 @@ def get_fastgeodis_func(num_dims, spacing=[1.0, 1.0, 1.0]):
     else:
         raise ValueError("Unsupported num_dims received: {}".format(num_dims))
 
+
 def get_signed_fastgeodis_func(num_dims, spacing=[1.0, 1.0, 1.0]):
     if num_dims == 2:
         return fastgeodis_signed_generalised_geodesic_distance_2d
     elif num_dims == 3:
-        return partial(fastgeodis_signed_generalised_geodesic_distance_3d, spacing=spacing)
+        return partial(
+            fastgeodis_signed_generalised_geodesic_distance_3d, spacing=spacing
+        )
     else:
         raise ValueError("Unsupported num_dims received: {}".format(num_dims))
+
 
 def get_toivanen_func(num_dims, spacing=[1.0, 1.0, 1.0]):
     if num_dims == 2:
@@ -160,9 +170,12 @@ def get_signed_toivanen_func(num_dims, spacing=[1.0, 1.0, 1.0]):
     if num_dims == 2:
[…]
