### 2022-07-23 8f0e0db
update benchmarking code
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
