### 2022-07-23 masadcv/FastGeodis#14 @masadcv
Apologies as this was automatically closed, reopening.

As mentioned in my comment above, I have implemented Toivanen's Geodesic method within FastGeodis package. This is a comparison existing algorithm that was compared to with in  GeodisTK. We dont have any dependency on GeodisTK anymore.

Can I please ask you to update to rc7 version and trying the failing samples again?
If all looks good, you can go ahead close this issue. Otherwise, please let me know if you see anything missing/not working.
