### 2022-07-23 c47c362
update readme with list of functions implemented
