+++ README.md
@@ -27,7 +27,7 @@ If you use this code in your research, then please consider citing:
 
 ***Asad, Muhammad, Reuben Dorent, and Tom Vercauteren. "FastGeodis: Fast Generalised Geodesic Distance Transform." Under review at Journal of Open Source Software (JOSS), 2022.***
 
-## Installation instructions
+# Installation instructions
 The provided package can be installed using:
 
 `pip install FastGeodis`
@@ -40,7 +40,30 @@ or (on conda systems with existing installation of PyTorch with CUDA)
 
 `pip install FastGeodis --no-build-isolation`
 
-## Example usage
+# Included methods
+## Optimised Fast Implementations for GPU/CPU based on [1, 3]
+| Method | Description | Documentation |
+|--------|-------------|---------------|
+| Fast Generalised Geodesic Distance 2D   |  Paralellised generalised geodesic distance transform for CPU/GPU [1, 3]          |      [FastGeodis.generalised_geodesic2d](https://masadcv.github.io/FastGeodis/#FastGeodis.generalised_geodesic2d)         |
+| Fast Generalised Geodesic Distance 3D   |  Paralellised generalised geodesic distance transform for CPU/GPU [1, 3]          |      [FastGeodis.generalised_geodesic3d](https://masadcv.github.io/FastGeodis/#FastGeodis.generalised_geodesic3d)         |
+| Fast Signed Generalised Geodesic Distance 2D   |  Paralellised signed generalised geodesic distance transform for CPU/GPU [1, 3]          |      [FastGeodis.signed_generalised_geodesic2d](https://masadcv.github.io/FastGeodis/#FastGeodis.signed_generalised_geodesic2d)         |
+| Fast Signed Generalised Geodesic Distance 3D   |  Paralellised signed generalised geodesic distance transform for CPU/GPU [1, 3]          |      [FastGeodis.signed_generalised_geodesic3d](https://masadcv.github.io/FastGeodis/#FastGeodis.signed_generalised_geodesic3d)         |
+| Fast Geodesic Symmetric Filtering 2D   |  Paralellised geodesic symmetric filtering for CPU/GPU [2]          |      [FastGeodis.GSF2d](https://masadcv.github.io/FastGeodis/#FastGeodis.GSF2d)         |
+| Fast Geodesic Symmetric Filtering 3D   |  Paralellised geodesic symmetric filtering for CPU/GPU [2]          |      [FastGeodis.GSF3d](https://masadcv.github.io/FastGeodis/#FastGeodis.GSF3d)         |
+----
+
+## Toivanen's Implementations for CPU based on [9]
+
+| Method | Description | Documentation |
+|--------|-------------|---------------|
+| Toivanen's Generalised Geodesic Distance 2D   |  Toivanen's generalised geodesic distance transform for CPU [9]          |      [FastGeodis.generalised_geodesic2d_toivanen](https://masadcv.github.io/FastGeodis/#FastGeodis.generalised_geodesic2d_toivanen)         |
+| Toivanen's Generalised Geodesic Distance 3D   |  Toivanen's generalised geodesic distance transform for CPU [9]          |      [FastGeodis.generalised_geodesic3d_toivanen](https://masadcv.github.io/FastGeodis/#FastGeodis.generalised_geodesic3d_toivanen)         |
[…]
