### 2022-07-23 #4532 @yhtang
I believe there are more recent works that handle the dependencies in a more sophisticated way. For example, see this [GTC talk](https://on-demand.gputechconf.com/gtc/2013/presentations/S3090-GPUGeoS-Geodesic-Distance-Transforms-for-Editing.pdf).
