### 2022-07-23 a1f09a5
add simpledemo3d, code formating
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
