+++ samples/test_speed_benchmark.py
@@ -1,11 +1,13 @@
-import numpy as np
-import torch
-import matplotlib.pyplot as plt
-from functools import wraps
-import SimpleITK as sitk
-import FastGeodis
-import time
 import os
+import time
+from functools import wraps
+
+import FastGeodis
+import matplotlib.pyplot as plt
+import numpy as np
+import SimpleITK as sitk
+import torch
+
 
 def timing(f):
     @wraps(f)
@@ -23,13 +25,16 @@ def timing(f):
 def generalised_geodesic_distance_2d_cpu(I, S, v, lamb, iter):
     return FastGeodis.generalised_geodesic2d_toivanen(I, S, v, lamb, iter)
 
+
 @timing
 def generalised_geodesic2d_raster_cpu(I, S, v, lamb, iter):
     return FastGeodis.generalised_geodesic2d(I, S, v, lamb, iter)
 
+
 @timing
 def generalised_geodesic2d_raster_gpu(I, S, v, lamb, iter):
     return FastGeodis.generalised_geodesic2d(I, S, v, lamb, iter)
+
 
 @timing
 def generalised_geodesic_distance_3d_cpu(I, S, spacing, v, lamb, iter):
@@ -40,17 +45,35 @@ def generalised_geodesic_distance_3d_cpu(I, S, spacing, v, lamb, iter):
 def generalised_geodesic3d_raster_cpu(I, S, spacing, v, lamb, iter):
     return FastGeodis.generalised_geodesic3d(I, S, spacing, v, lamb, iter)
 
+
 @timing
 def generalised_geodesic3d_raster_gpu(I, S, spacing, v, lamb, iter):
     return FastGeodis.generalised_geodesic3d(I, S, spacing, v, lamb, iter)
-    
-func_to_test_2d = [generalised_geodesic_distance_2d_cpu, generalised_geodesic2d_raster_cpu, generalised_geodesic2d_raster_gpu]
-func_to_test_3d = [generalised_geodesic_distance_3d_cpu, generalised_geodesic3d_raster_cpu, generalised_geodesic3d_raster_gpu]
+
+
+func_to_test_2d = [
+    generalised_geodesic_distance_2d_cpu,
+    generalised_geodesic2d_raster_cpu,
+    generalised_geodesic2d_raster_gpu,
+]
+func_to_test_3d = [
+    generalised_geodesic_distance_3d_cpu,
+    generalised_geodesic3d_raster_cpu,
+    generalised_geodesic3d_raster_gpu,
+]
+
 
 def test2d():
     num_runs = 5
 
-    sizes_to_test = [64, 64*2, 64*(2**2), 64*(2**3), 64*(2**4), 64*(2**5)]#, 64*(2**6)]
+    sizes_to_test = [
+        64,
+        64 * 2,
+        64 * (2 ** 2),
+        64 * (2 ** 3),
+        64 * (2 ** 4),
+        64 * (2 ** 5),
+    ]  # , 64*(2**6)]
     print(sizes_to_test)
     time_taken_dict = dict()
     for func in func_to_test_2d:
@@ -58,31 +81,43 @@ def test2d():
         for size in sizes_to_test:
             image = torch.rand((1, 1, size, size))
             seed = torch.ones((1, 1, size, size))
-            seed[:, :, size//2, size//2] = 0.0
+            seed[:, :, size // 2, size // 2] = 0.0
 
-            if 'gpu' in func.__name__:
-                image = image.to('cuda').contiguous()
-                seed = seed.to('cuda').contiguous()
-            
+            if "gpu" in func.__name__:
+                image = image.to("cuda").contiguous()
+                seed = seed.to("cuda").contiguous()
+
             tic = time.time()
             for i in range(num_runs):
[…]
