+++ samples/simpledemo2d_signed.py
@@ -1,9 +1,8 @@
+import FastGeodis
 import matplotlib.pyplot as plt
 import numpy as np
 import torch
 from PIL import Image
-
-import FastGeodis
 
 device = "cuda" if torch.cuda.is_available() else "cpu"
 image = np.asarray(Image.open("../data/img2d.png"), np.float32)
@@ -16,13 +15,13 @@ mask_pt[..., 100, 100] = 0
 v = 1e10
 iterations = 2
 
-lamb = 1.0 # <-- Geodesic distance transform
+lamb = 1.0  # <-- Geodesic distance transform
 geodesic_dist = FastGeodis.signed_generalised_geodesic2d(
     image_pt, mask_pt, v, lamb, iterations
 )
 geodesic_dist = np.squeeze(geodesic_dist.cpu().numpy())
 
-lamb=0.0 # <-- Euclidean distance transform
+lamb = 0.0  # <-- Euclidean distance transform
 euclidean_dist = FastGeodis.signed_generalised_geodesic2d(
     image_pt, mask_pt, v, lamb, iterations
 )
@@ -52,16 +51,16 @@ print(euclidean_dist.max())
 # Toivanen's Raster Method
 
 # Toivanen's method only support CPU
-image_pt = image_pt.to('cpu')
-mask_pt = mask_pt.to('cpu')
+image_pt = image_pt.to("cpu")
+mask_pt = mask_pt.to("cpu")
 
-lamb = 1.0 # <-- Geodesic distance transform
+lamb = 1.0  # <-- Geodesic distance transform
 geodesic_dist_toivanen = FastGeodis.signed_generalised_geodesic2d_toivanen(
     image_pt, mask_pt, v, lamb, iterations
 )
 geodesic_dist_toivanen = np.squeeze(geodesic_dist_toivanen.cpu().numpy())
 
-lamb=0.0 # <-- Euclidean distance transform
+lamb = 0.0  # <-- Euclidean distance transform
 euclidean_dist_toivanen = FastGeodis.signed_generalised_geodesic2d_toivanen(
     image_pt, mask_pt, v, lamb, iterations
 )

[…]
