### 2022-07-23 2f3f8f8
update simple samples
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
