+++ samples/simpledemo2d_signed.py
@@ -5,7 +5,7 @@ import torch
 from PIL import Image
 
 device = "cuda" if torch.cuda.is_available() else "cpu"
-image = np.asarray(Image.open("../data/img2d.png"), np.float32)
+image = np.asarray(Image.open("data/img2d.png"), np.float32)
 
 image_pt = torch.from_numpy(image).unsqueeze_(0).unsqueeze_(0)
 image_pt = image_pt.to(device)
@@ -27,7 +27,7 @@ euclidean_dist = FastGeodis.signed_generalised_geodesic2d(
 )
 euclidean_dist = np.squeeze(euclidean_dist.cpu().numpy())
 
-plt.figure(figsize=(16, 12))
+plt.figure(figsize=(12, 4))
 plt.subplot(1, 3, 1)
 plt.imshow(image)
 
@@ -66,7 +66,7 @@ euclidean_dist_toivanen = FastGeodis.signed_generalised_geodesic2d_toivanen(
 )
 euclidean_dist_toivanen = np.squeeze(euclidean_dist_toivanen.cpu().numpy())
 
-plt.figure(figsize=(16, 12))
+plt.figure(figsize=(12, 4))
 plt.subplot(1, 3, 1)
 plt.imshow(image)
 
