+++ samples/simpledemo2d.py
@@ -7,7 +7,7 @@ from PIL import Image
 # FastGeodis Method
 
 device = "cuda" if torch.cuda.is_available() else "cpu"
-image = np.asarray(Image.open("../data/img2d.png"), np.float32)
+image = np.asarray(Image.open("data/img2d.png"), np.float32)
 
 image_pt = torch.from_numpy(image).unsqueeze_(0).unsqueeze_(0)
 image_pt = image_pt.to(device)
@@ -29,7 +29,7 @@ euclidean_dist = FastGeodis.generalised_geodesic2d(
 )
 euclidean_dist = np.squeeze(euclidean_dist.cpu().numpy())
 
-plt.figure(figsize=(16, 12))
+plt.figure(figsize=(12, 4))
 plt.subplot(1, 3, 1)
 plt.imshow(image)
 
@@ -50,18 +50,18 @@ image_pt = image_pt.to("cpu")
 mask_pt = mask_pt.to("cpu")
 
 lamb = 1.0  # <-- Geodesic distance transform
-geodesic_dist_toivanen = FastGeodis.signed_generalised_geodesic2d_toivanen(
+geodesic_dist_toivanen = FastGeodis.generalised_geodesic2d_toivanen(
     image_pt, mask_pt, v, lamb, iterations
 )
 geodesic_dist_toivanen = np.squeeze(geodesic_dist_toivanen.cpu().numpy())
 
 lamb = 0.0  # <-- Euclidean distance transform
-euclidean_dist_toivanen = FastGeodis.signed_generalised_geodesic2d_toivanen(
+euclidean_dist_toivanen = FastGeodis.generalised_geodesic2d_toivanen(
     image_pt, mask_pt, v, lamb, iterations
 )
 euclidean_dist_toivanen = np.squeeze(euclidean_dist_toivanen.cpu().numpy())
 
-plt.figure(figsize=(16, 12))
+plt.figure(figsize=(12, 4))
 plt.subplot(1, 3, 1)
 plt.imshow(image)
 
