### 2022-07-25 #4532 @masadcv
> I believe there are more recent works that handle the dependencies in a more sophisticated way. For example, see this [GTC talk](https://on-demand.gputechconf.com/gtc/2013/presentations/S3090-GPUGeoS-Geodesic-Distance-Transforms-for-Editing.pdf).

Many thanks for sharing this @yhtang !

We are already aware of Weber et al's [1] method (we already cite this in our paper), this method is protected by multiple patents [2, 3, 4[ and hence we did not spent time implementing it.

[1] Weber, Ofir, et al. "Parallel algorithms for approximation of distance maps on parametric surfaces." ACM Transactions on Graphics (TOG), (2008).
[2] Bronstein, Alexander, et al. "Parallel approximation of distance maps." U.S. Patent No. 8,373,716. 12 Feb. 2013.
[3] Bronstein, Alexander, et al. "Parallel approximation of distance maps." U.S. Patent No. 8,982,142. 17 Mar. 2015.
[4] Bronstein, Alexander, et al. "Parallel approximation of distance maps." U.S. Patent No. 9,489,708. 8 Nov. 2016.
