### 2022-07-25 masadcv/FastGeodis#20 @masadcv
[BUG] Improve CUDA kernels to reduce number of invocations
As reported in https://github.com/openjournals/joss-reviews/issues/4532 in the comment https://github.com/openjournals/joss-reviews/issues/4532#issuecomment-1192949580

The cuda kernels are called many times (invocations == number of rows/planes). As indicated by reviewer, this may incur additional costs for setting up and running kernels each time on new row.

An expected behviour here would be to setup and call kernel for each pass once. This will reduce the overhead added by many calls and would potentially improve the overall execution efficiency of the current algorithms.
