### 2022-07-25 #4532 @masadcv
> More interestingly, I just ran a simplified version of the [Simple 2D Geodesic and Euclidean Distance](https://colab.research.google.com/github/masadcv/FastGeodis/blob/master/samples/simpledemo2d.ipynb) example as follow and found that the `geodesic_updown_single_row_pass_kernel` kernel gets launched by **1336 times**, each time with a 1x1x1 grid and 256x1x1 block. This could be easily converted to a single launch with a 1336x1x1 grid and 256x1x1 block size to avoid the huge kernel launch overhead currently invoked. This is further evidence that the software in its current form lacks a certain amount of maturity.

@yhtang I have opened an issue related to this and should have a PR to reduce the number of invocations in the current algorithm soon.
