### 2022-07-25 05b834c
adding improvements to reduce num kernel calls
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
