+++ FastGeodis/fastgeodis_cuda.cu
@@ -57,54 +57,72 @@ __global__ void geodesic_updown_single_row_pass_kernel(
     torch::PackedTensorAccessor32<scalar_t, 4, torch::RestrictPtrTraits> distance_ptr,
     const float l_grad,
     const float l_eucl,
-    const int direction,
-    const int h)
+    const int direction)
 {
     const int channel = image_ptr.size(1);
+    const int height = image_ptr.size(2);
     const int width = image_ptr.size(3);
 
     int kernelW = blockIdx.x * blockDim.x + threadIdx.x;
 
+    int h = (direction == 1)? 1 : height - 2; 
+
     // if outside, then skip distance calculation - dont use the thread
     if (kernelW < width)
     {
-        float l_dist, cur_dist;
-        float new_dist = distance_ptr[0][0][h][kernelW];
-
-        for (int w_i = 0; w_i < 3; w_i++)
+        while (h >= 0 && h < height)
         {
-            const int kernelW_ind = kernelW + w_i - 1;
-
-            if (kernelW_ind >= 0 && kernelW_ind < width)
+            int prevH = h - direction;
+            if (prevH < 0 || prevH >= height)
             {
-                l_dist = 0.0;
-                if (channel == 1)
+                // read outside bounds, skip
+                continue;
+            }
+            
+            float l_dist, cur_dist;
+            float new_dist = distance_ptr[0][0][h][kernelW];
+
+            for (int w_i = 0; w_i < 3; w_i++)
+            {
+                const int kernelW_ind = kernelW + w_i - 1;
+
+                if (kernelW_ind >= 0 && kernelW_ind < width)
                 {
-                    l_dist = l1distance_cuda(
-                        image_ptr[0][0][h][kernelW], 
-                        image_ptr[0][0][h + direction][kernelW_ind]
-                        );
-                }
-                else
-                {
-                    for (int c_i = 0; c_i < channel; c_i++)
+                    l_dist = 0.0;
+                    if (channel == 1)
                     {
-                        l_dist += l1distance_cuda(
-                            image_ptr[0][c_i][h][kernelW], 
-                            image_ptr[0][c_i][h + direction][kernelW_ind]
+                        l_dist = l1distance_cuda(
+                            image_ptr[0][0][h][kernelW], 
+                            image_ptr[0][0][prevH][kernelW_ind]
                             );
                     }
-                }
-                cur_dist = distance_ptr[0][0][h + direction][kernelW_ind] + \
-                            l_eucl * local_dist2d[w_i] + \
-                            l_grad * l_dist;
+                    else
+                    {
+                        for (int c_i = 0; c_i < channel; c_i++)
+                        {
+                            l_dist += l1distance_cuda(
+                                image_ptr[0][c_i][h][kernelW], 
+                                image_ptr[0][c_i][prevH][kernelW_ind]
+                                );
+                        }
+                    }
+                    cur_dist = distance_ptr[0][0][prevH][kernelW_ind] + \
+                                l_eucl * local_dist2d[w_i] + \
+                                l_grad * l_dist;
 
-                new_dist = std::min(new_dist, cur_dist);
+                    new_dist = std::min(new_dist, cur_dist);
+                }
             }
-        }
-        if (new_dist < distance_ptr[0][0][h][kernelW])
-        {
-            distance_ptr[0][0][h][kernelW] = new_dist;
+            if (new_dist < distance_ptr[0][0][h][kernelW])
+            {
+                distance_ptr[0][0][h][kernelW] = new_dist;
+            }
+
+            // go to next row
+            h += direction;
+
+            // synchronise write for all threads
+            __syncthreads();
         }
[…]
