### 2022-07-25 6e14769
improve use_ptr code
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
