### 2022-07-25 #4532 @yhtang
> this method is protected by multiple patents

Thanks for sharing this knowledge. This is sad and ridiculous. But since the NVIDIA guy already implemented and presented it, I suppose maybe a better understanding of the scope of the patent could lead to workarounds?

Regardless of the patent, there are still many commonly used CUDA techniques that can be potentially used to prevent the kernels in FastGeodis from being memory-bound (I don't think anyone can patent these). For example, you can trade computation for dependency: store multiple rows of data in shared memory and then let each thread compute its dependent pixel values from scratch. This way, some computations are duplicated among the threads. But since FLOPS are nearly free on a GPU, we can still expect an overall speedup due to less memory traffic and fewer kernel launches. Have you considered this option yet?
