### 2022-07-25 #4532 @masadcv
> Regardless of the patent, there are still many commonly used CUDA techniques that can be potentially used to prevent the kernels in FastGeodis from being memory-bound (I don't think anyone can patent these). For example, you can trade computation for dependency: store multiple rows of data in shared memory and then let each thread compute its dependent pixel values from scratch. This way, some computations are duplicated among the threads. But since FLOPS are nearly free on a GPU, we can still expect an overall speedup due to less memory traffic and fewer kernel launches. Have you considered this option yet?

@yhtang Great idea, I had not thought of this before. I have just now implemented something similar to address the invocation issue that you reported above. this PR https://github.com/masadcv/FastGeodis/pull/21 shows that indeed having less kernel calls leads to improvement in overall performance. The threads now work on calculation on their calculation for a given row, and synchronise to after each row...
I will look into shared memory and see if I can work out the trade-off between compute vs dependency, as you mentioned.
