### 2022-07-25 #4532 @masadcv
Tracking paper changes requested by @jacobmerson in [this comment above](https://github.com/openjournals/joss-reviews/issues/4532#issuecomment-1191085210)

- [ ] What is Geodesic distance? How it differs from Euclidean distance? What is segmentation?
- [x] Provide the hardware information for speedup claims.
- [ ] Be clear about what you are comparing, CPU single thread vs GPU? or something else.
- [ ] Is your performance relevant to 2D or 3D?
- [ ] In summary section, give at least a hint as to why you are seeing better performance?
- [ ] Clear description of what each pass is, and how it is parallelised on CPU/GPU.

> Based on the Joss guidelines, it’s not clear that the methodology section is needed in the paper.

@vissarion may I ask for guidance on the above point? Can we have a methodology section in paper that includes equations and description of methods? As far as I understood from [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), JOSS papers may not need to detail the actual methods/API/equations... etc . If so, I will add this detail to documentation instead.
