### 2022-07-26 #4532 @vissarion
> @vissarion may I ask for guidance on the above point? Can we have a methodology section in paper that includes equations and description of methods? As far as I understood from [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), JOSS papers may not need to detail the actual methods/API/equations... etc . If so, I will add this detail to documentation instead.

According to JOSS guidelines novel results should not be placed in the paper. That is, if there are (research) papers describing the method implemented by the package you should site those. Adding a detailed description in the documentation and provide the link in the paper is ideal. Please also add (in the same section in the docs) the explanations in your comment above to one of the reviewers (@yhtang). It seems that those explanations added some missing clarity.
