### 2022-07-27 8437f04
adding exact dependency versions
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
