### 2022-07-27 9e95136
adding better docu template, filled, only methodology remains
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
