+++ docs/source/conf.py
@@ -34,6 +34,7 @@ extensions = [
         'sphinx.ext.napoleon',
         'sphinx.ext.viewcode',
         'sphinx_rtd_theme',
+        'sphinxemoji.sphinxemoji',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
