+++ docs/source/contributing.rst
@@ -1,0 +1,24 @@
+*****************
+Contributing
+*****************
+
+
+Bugs |:bug:|
+################################
+If you found a bug in FastGeodis, please `open an issue request in GitHub. <https://github.com/masadcv/FastGeodis/issues>`_ Please be as descriptive as possible in your issue request.
+
+Feature suggestions |:bulb:|
+################################
+If you'd like to suggest a feature to be added to FastGeodis, please `open an issue request in GitHub. <https://github.com/masadcv/FastGeodis/issues>`_ and describe your feature appropriately.
+
+
+Contributing |:computer:|
+################################
+The following steps should be taken to contribute to FastGeodis:
+
+1. Fork the FastGeodis repository.
+2. Create a branch for any changes you are contributing and add your changes locally.
+3. Ensure that unittests pass locally for you changes.
+4. Commit your code to the branch in your fork.
+5. Create a pull request that documents your additions/changes.
+6. Add as much description for your proposed changes and fix any build issues within Github CI.
