+++ docs/_sources/methodology.rst.txt
@@ -1,0 +1,35 @@
+*****************
+Methodology
+*****************
+
+Fundamental Concepts
+################################
+
+
+Generalised Geodesic Distance Transform
+***************************************
+
+
+Interactive Segmentation using Geodesic Distance Transform
+**********************************************************
+
+
+
+
+Implemented Algorithms
+################################
+
+Criminisi et al's Parallelisable Generalised Geodesic Distance Transform
+************************************************************************
+
+Toivanen et al's Non-parallelisable Generalised Geodesic Distance Transform
+***************************************************************************
+
+
+
+
+Performance Improvements
+################################
+
+
+
