+++ docs/Makefile
@@ -19,4 +19,4 @@ help:
 # "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
 %: Makefile
 	@$(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
-	cp -r $(BUILDDIR)/singlehtml/* $(HTMLCOPYDIR)
+	cp -r $(BUILDDIR)/html/* $(HTMLCOPYDIR)
