### 2022-07-27 9c7cf9c
add contribution guide section in readmemd
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
