+++ docs/requirements-doc.txt
@@ -1,0 +1,1 @@
+sphinxemoji
