+++ docs/requirements-doc.txt
@@ -1,1 +1,2 @@
 sphinxemoji
+torch
