+++ docs/source/conf.py
@@ -12,7 +12,7 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../'))
+sys.path.insert(0, os.path.abspath('../../FastGeodis'))
 
 
 # -- Project information -----------------------------------------------------
@@ -54,6 +54,28 @@ exclude_patterns = []
 html_theme = 'sphinx_rtd_theme'
 
 # Add any paths that contain custom static files (such as style sheets) here,
+
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
+# help from: https://github.com/PyLops/pylops/blob/dev/docs/source/conf.py
+html_context = {
+    "menu_links_name": "Repository",
+    "menu_links": [
+        (
+            '<i class="fa fa-github fa-fw"></i> Source Code',
+            "https://github.com/masadcv/FastGeodis",
+        ),
+        (
+            '<i class="fa fa-users fa-fw"></i> Contributing',
+            "https://github.com/masadcv/FastGeodis/blob/master/CONTRIBUTING.md",
+        ),
+    ],
+    # Custom variables to enable "Improve this page"" and "Download notebook"
+    # links
+    "doc_path": "docs/source",
+    "github_project": "masadcv",
+    "github_repo": "FastGeodis",
+    "github_version": "master",
+}
