+++ docs/requirements-doc.txt
@@ -1,2 +1,1 @@
 sphinxemoji
-torch
