+++ docs/source/citing.rst
@@ -1,0 +1,6 @@
+**************************
+How to Cite FastGeodis
+**************************
+
+When using **FastGeodis** in scientific publications, please cite the following paper:
+
