### 2022-07-27 6428854
move docs to readthedocs hosting
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
