### 2022-07-27 0990b51
adding methodology documentation, evaluation, citations in docs
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
