+++ samples/test_speed_benchmark_geodistk.py
@@ -1,0 +1,141 @@
+import json
+import GeodisTK
+import numpy as np
+import torch
+import matplotlib.pyplot as plt
+from functools import wraps
+import SimpleITK as sitk
+import FastGeodis
+import time
+import os
+
+def timing(f):
+    @wraps(f)
+    def wrap(*args, **kw):
+        ts = time.time()
+        result = f(*args, **kw)
+        te = time.time()
+        print("func:%r took: %2.4f sec" % (f.__name__, te - ts))
+        return result
+
+    return wrap
+
+@timing
+def generalised_geodesic_distance_2d(I, S, v, lamb, iter):
+    return GeodisTK.geodesic2d_raster_scan(I, 1-S.astype(np.uint8), lamb, iter)
+
+@timing
+def generalised_geodesic2d_raster_cpu(I, S, v, lamb, iter):
+    return FastGeodis.generalised_geodesic2d(I, S, v, lamb, iter)
+
+@timing
+def generalised_geodesic2d_raster_gpu(I, S, v, lamb, iter):
+    return FastGeodis.generalised_geodesic2d(I, S, v, lamb, iter)
+
+@timing
+def generalised_geodesic_distance_3d(I, S, spacing, v, lamb, iter):
+    return GeodisTK.geodesic3d_raster_scan(I, 1-S.astype(np.uint8), spacing, lamb, iter)
+
+@timing
+def generalised_geodesic3d_raster_cpu(I, S, spacing, v, lamb, iter):
+    return FastGeodis.generalised_geodesic3d(I, S, spacing, v, lamb, iter)
+
+@timing
+def generalised_geodesic3d_raster_gpu(I, S, spacing, v, lamb, iter):
+    return FastGeodis.generalised_geodesic3d(I, S, spacing, v, lamb, iter)
+    
+func_to_test_2d = [generalised_geodesic_distance_2d, generalised_geodesic2d_raster_cpu, generalised_geodesic2d_raster_gpu]
+func_to_test_3d = [generalised_geodesic_distance_3d, generalised_geodesic3d_raster_cpu, generalised_geodesic3d_raster_gpu]
+
+def test2d():
+    num_runs = 10
+
+    sizes_to_test = [64, 64*2, 64*(2**2), 64*(2**3), 64*(2**4), 64*(2**5)]#, 64*(2**6)]
+    print(sizes_to_test)
+    time_taken_dict = dict()
+    for func in func_to_test_2d:
+        time_taken_dict[func.__name__] = []
+        for size in sizes_to_test:
+            image = torch.rand((1, 1, size, size))
+            seed = torch.ones((1, 1, size, size))
+            seed[:, :, size//2, size//2] = 0.0
+
+            if 'gpu' in func.__name__:
+                image = image.to('cuda').contiguous()
+                seed = seed.to('cuda').contiguous()
+            
+            tic = time.time()
+            for i in range(num_runs):
+                if 'cpu' in func.__name__:
+                    func(image, seed, 10000, 1.0, 2)
+                elif 'gpu' in func.__name__ and torch.cuda.is_available():
+                    func(image, seed, 10000, 1.0, 2)
+                else:
+                    func(np.squeeze(image.cpu().numpy()), np.squeeze(seed.cpu().numpy()), 10000, 1.0, 2)
+
+            time_taken_dict[func.__name__].append((time.time() - tic)/num_runs)
+        print()
+
+    return sizes_to_test, time_taken_dict
+
+def test3d():
+    num_runs = 10
+    spacing = [1.0, 1.0, 1.0]
+
[…]
