+++ docs/source/methodology.rst
@@ -2,34 +2,99 @@
 Methodology
 *****************
 
-Fundamental Concepts
-################################
-
-
 Generalised Geodesic Distance Transform
-***************************************
+#######################################
 
+A distance transform maps each image pixel/voxel into its smallest distance to regions of interest :cite:p:`fabbri20082d`. It is a fundamental operator with relevant applications in computer vision, graphics, shape analysis, pattern recognition and computational geometry :cite:p:`fabbri20082d`. 
 
-Interactive Segmentation using Geodesic Distance Transform
-**********************************************************
+Generalised Geodesic distance transform combines spatial as well as gradient based distance transform and is defined as :cite:p:`criminisi2008geos`:
 
+.. math:: 
+    d(\mathbf{x}; I, \lambda) = \min \sum_{k=1}^{n} \Big[ \underbrace{\left( 1-\lambda \right) \left \| \mathbf{x}_k - \mathbf{x}_{k-1}  \right \|_{L1}}_{\text{Euclidean Term}} + \underbrace{\lambda \left | I(\mathbf{x}_k) - I(\mathbf{x}_{k-1})  \right |}_{\text{Geodesic Term}} \Big]
 
+The ability to incorporate image gradients with spatial distances has enabled application of Geodesic distance transforms in a number of areas, including image editing and filtering :cite:p:`criminisi2008geos`.  Methods from recent years have shown effectiveness in applying Geodesic distance transform to interactively annotate 3D medical imaging data :cite:p:`wang2018deepigeos,criminisi2008geos`, where it enables providing segmentation labels, i.e. voxel-wise labels, for different objects of interests.
 
+Statement of Need
+##################
 
-Implemented Algorithms
-################################
+Despite existing open-source implementation of distance transforms :cite:p:`tensorflow2015-whitepaper,eucildeantdimpl,geodistk`, open-source implementations of efficient Geodesic distance transform algorithms :cite:p:`criminisiinteractive,weber2008parallel` on CPU and GPU do not exist. However, efficient CPU :cite:p:`eucildeantdimpl` and GPU :cite:p:`tensorflow2015-whitepaper` implementations exist for Euclidean distance transform. To the best of our knowledge, **FastGeodis** is the first open-source implementation of efficient Geodesic distance transform :cite:p:`criminisiinteractive`, achieving up to 15x speed-up on CPU and up to 60x speed-up on GPU as compared to existing open-source libraries :cite:p:`geodistk`. It also provides efficient implementation of Euclidean distance transform. In addition, it is the first open-source implementation of generalised Geodesic distance transform and Geodesic Symmetric Filtering (GSF) proposed in :cite:p:`criminisi2008geos`. 
+  
+
+The ability to efficiently compute Geodesic and Euclidean distance transforms can significantly enhance distance transform applications especially for training deep learning models that utilise distance transforms :cite:p:`wang2018deepigeos`. It will improve prototyping, experimentation, and deployment of such methods, where efficient computation of distance transforms has been a limiting factor. In 3D medical imaging problems, efficient computation of distance transforms will lead to significant speed-ups, enabling online learning applications for better processing/labelling/inference from volumetric datasets :cite:p:`asad2022econet`.  In addition, **FastGeodis** provides efficient implementation for both CPUs and GPUs hardware and hence will enable efficient use of a wide range of hardware devices. 
+
+Implementation
+#################
[…]
