### 2022-07-27 c95d3be
add json results for performance evaluation
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
