+++ samples/test_speed_benchmark_toivanen.py
@@ -171,7 +171,7 @@ def save_plot(sizes, time_taken_dict, figname):
     plt.ylabel("Execution time (seconds)")
     plt.tight_layout()
     plt.savefig(os.path.join("figures", figname + ".png"))
-    time_taken_dict['xaxis'] = sizes
+    time_taken_dict['spatial_dim'] = sizes
     with open(os.path.join('figures', figname + '.json'), 'w') as fp:
         json.dump(time_taken_dict, fp, indent=4)
 
