+++ docs/source/acknowledgement.rst
@@ -1,6 +1,7 @@
 **************************
-Acknowledgment
+Acknowledgement
 **************************
 
 Development of **FastGeodis** was supported by the European Union's Horizon 2020 research and innovation programme under grant agreement No 101016131. 
 
+We would also like to thank reviewers and editors for the `Journal of Open Source Software`<https://joss.theoj.org/>_ for their valuable feedback and comments to further improve this package.
