### 2022-07-27 9297c7e
adding updated methodology section
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
