### 2022-07-27 #4532 @masadcv
Just an update from our side.

We have taken all comments provided so far and prepared detailed documentation that covers different aspects of the package.
This includes paper comments from @jacobmerson , execution speed comparison and algorithmic detail comments from @yhtang  and instructions from @vissarion on adding details from above comments into documentation.

This can be accessed and reviewed at: https://fastgeodis.readthedocs.io/en/latest/index.html
In particular, you may be interested in reviewing the methodology section found here: https://fastgeodis.readthedocs.io/en/latest/methodology.html

I have also updated the paper with new numbers generated after merging PR (https://github.com/masadcv/FastGeodis/pull/21) which was based on comments from @yhtang to minimize the kernel invocations (https://github.com/openjournals/joss-reviews/issues/4532#issuecomment-1192949580)

We welcome any feedback, comments/suggestion on this.
