### 2022-07-27 f3b4cb4
fix joss link in doc
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
