### 2022-07-28 #4532 @vissarion
We are currently waiting for the scope check by JOSS editorial team. Thia could take some time since this is a vacation period for many places on Earth.

@masadcv thanks for the feedback and all the work!

@jacobmerson and @yhtang since the mentality of JOSS is not rejecting outright submissions requiring major revisions but rather giving the author as long as they need to make modifications/improvements, is it possible for you to state the minimum requirements that you as a reviewer need to accept this submission.
