### 2022-08-01 #4532 @jacobmerson
@masadcv I just returned from vacation. I will take a look at the changes that have been made shortly. Thanks for making a checklist of my comments.

Based on the discussion with @yhtang I think you should include citations of the patents you list in this thread and the nvidia talk. This is important context for why you have implemented the chosen algorithms.
