+++ README.md
@@ -25,7 +25,17 @@ If you use this code in your research, then please consider citing:
 
 [![status](https://joss.theoj.org/papers/d0b6e3daa4b22fec471691c6f1c60e2a/status.svg)](https://joss.theoj.org/papers/d0b6e3daa4b22fec471691c6f1c60e2a)
 
-***Asad, Muhammad, Reuben Dorent, and Tom Vercauteren. "FastGeodis: Fast Generalised Geodesic Distance Transform." Under review at Journal of Open Source Software (JOSS), 2022.***
+***Asad, Muhammad, Reuben Dorent, and Tom Vercauteren. "FastGeodis: Fast Generalised Geodesic Distance Transform." Under review at Journal of Open Source Software (JOSS), 2022.*** ([preprint](https://arxiv.org/pdf/2208.00001.pdf))
+
+Bibtex:
+```
+@article{asad2022fastgeodis,
+  title={FastGeodis: Fast Generalised Geodesic Distance Transform},
+  author={Asad, Muhammad and Dorent, Reuben and Vercauteren, Tom},
+  journal={arXiv preprint arXiv:2208.00001},
+  year={2022}
+}
+```
 
 # Installation instructions
 The provided package can be installed using:
