+++ FastGeodis/geodis_fastmarch.cpp
@@ -1,0 +1,426 @@
+// BSD 3-Clause License
+
+// Copyright (c) 2021, Muhammad Asad (masadcv@gmail.com)
+// All rights reserved.
+
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+
+// 1. Redistributions of source code must retain the above copyright notice, this
+//    list of conditions and the following disclaimer.
+
+// 2. Redistributions in binary form must reproduce the above copyright notice,
+//    this list of conditions and the following disclaimer in the documentation
+//    and/or other materials provided with the distribution.
+
+// 3. Neither the name of the copyright holder nor the names of its
+//    contributors may be used to endorse or promote products derived from
+//    this software without specific prior written permission.
+
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+// DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+// CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+// OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+#include <torch/extension.h>
+#include <vector>
+// #include <iostream>
+
+float l1distance_toivanen(const float &in1, const float &in2)
+{
+    return std::abs(in1 - in2);
+}
+
+struct Point2D
+{
+    float distance;
+    int w;
+    int h;
+};
+
+void insert_point_to_list(std::vector<Point2D> * list, int start_position,  Point2D p)
+{
+    int insert_idx = list->size();
+    for(int i = start_position; i < list->size(); i++)
+    {
+        if(list->at(i).distance < p.distance)
+        {
+            insert_idx = i;
+            break;
+        }
+    }
+    list->insert(insert_idx + list->begin(), p);
+}
+
+void update_point_in_list(std::vector<Point2D> * list, Point2D p)
+{
+    int remove_idx = -1;
+    for(int i = 0; i < list->size(); i++)
+    {
+        if(list->at(i).w == p.w && list->at(i).h == p.h)
+        {
+            remove_idx = i;
+            break;
+        }
+    }
+    list->erase(remove_idx + list->begin());
+    insert_point_to_list(list, remove_idx, p);
+}
+
+void add_new_accepted_point(
+    const torch::TensorAccessor<float, 4> &img, 
+    torch::TensorAccessor<float, 4> &distance, 
+    torch::TensorAccessor<int, 2> &state, 
+    std::vector<Point2D> * list, 
+    int height, 
+    int width, 
+    int channel, 
+    Point2D p)
+{
+    int w = p.w;
+    int h = p.h;
[…]
