### 2022-08-14 b8ccc25
adding python interface for all FastMarch, working FastMarch2d, needs testing
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
