+++ FastGeodis/__init__.py
@@ -308,6 +308,145 @@ def signed_generalised_geodesic3d_toivanen(
         image, softmask, spacing, v, lamb, 1 - lamb, iter
     )
 
+def generalised_geodesic2d_fastmarch(
+    image: torch.Tensor, 
+    softmask: torch.Tensor, 
+    v: float, 
+    lamb: float
+):
+    r"""Computes Generalised Geodesic Distance using Fast Marching method from:
+
+    Sethian, James A. 
+    "Fast marching methods." 
+    SIAM review 41.2 (1999): 199-235.
+    
+    For more details on generalised geodesic distance, check the following reference:
+
+    Criminisi, Antonio, Toby Sharp, and Andrew Blake.
+    "Geos: Geodesic image segmentation."
+    European Conference on Computer Vision, Berlin, Heidelberg, 2008.
+
+    The function expects input as torch.Tensor, which can be run on CPU only using Tensor's device location
+
+    Args:
+        image: input image, can be grayscale or multiple channels.
+        softmask: softmask in range [0, 1] with seed information.
+        v: weighting factor for establishing relationship between unary and spatial distances.
+        lamb: weighting factor between 0.0 and 1.0. 0.0 returns euclidean distance, whereas 1.0 returns geodesic distance
+
+    Returns:
+        torch.Tensor with distance transform
+    """
+    return FastGeodisCpp.generalised_geodesic2d_fastmarch(
+        image, softmask, v, lamb, 1 - lamb
+    )
+
+
+def generalised_geodesic3d_fastmarch(
+    image: torch.Tensor,
+    softmask: torch.Tensor,
+    spacing: List,
+    v: float,
+    lamb: float
+):
+    r"""Computes Generalised Geodesic Distance using Fast Marching method from:
+
+    TSethian, James A. 
+    "Fast marching methods." 
+    SIAM review 41.2 (1999): 199-235.
+    
+    For more details on generalised geodesic distance, check the following reference:
+
+    Criminisi, Antonio, Toby Sharp, and Andrew Blake.
+    "Geos: Geodesic image segmentation."
+    European Conference on Computer Vision, Berlin, Heidelberg, 2008.
+
+    The function expects input as torch.Tensor, which can be run on CPU only using Tensor's device location
+
+    Args:
+        image: input image, can be grayscale or multiple channels.
+        softmask: softmask in range [0, 1] with seed information.
+        spacing: spacing for 3D data
+        v: weighting factor for establishing relationship between unary and spatial distances.
+        lamb: weighting factor between 0.0 and 1.0. 0.0 returns euclidean distance, whereas 1.0 returns geodesic distance
+
+    Returns:
+        torch.Tensor with distance transform
+    """
+    return FastGeodisCpp.generalised_geodesic3d_fastmarch(
+        image, softmask, spacing, v, lamb, 1 - lamb
+    )
+
+def signed_generalised_geodesic2d_fastmarch(
+    image: torch.Tensor, 
+    softmask: torch.Tensor, 
+    v: float, 
+    lamb: float
+):
+    r"""Computes Signed Generalised Geodesic Distance using Fast Marching method from:
+
+    Sethian, James A. 
+    "Fast marching methods." 
+    SIAM review 41.2 (1999): 199-235.
+    
+    For more details on generalised geodesic distance, check the following reference:
+
+    Criminisi, Antonio, Toby Sharp, and Andrew Blake.
+    "Geos: Geodesic image segmentation."
+    European Conference on Computer Vision, Berlin, Heidelberg, 2008.
+
+    The function expects input as torch.Tensor, which can be run on CPU only using Tensor's device location
+
+    Args:
[…]
