+++ FastGeodis/fastgeodis.h
@@ -87,6 +87,21 @@ torch::Tensor generalised_geodesic3d_toivanen_cpu(
     const float &l_eucl,
     const int &iterations);
 
+torch::Tensor generalised_geodesic2d_fastmarch_cpu(
+    torch::Tensor &image,
+    const torch::Tensor &mask,
+    const float &v,
+    const float &l_grad,
+    const float &l_eucl);
+
+// torch::Tensor generalised_geodesic3d_fastmarch_cpu(
+//     torch::Tensor &image,
+//     const torch::Tensor &mask,
+//     const std::vector<float> &spacing,
+//     const float &v,
+//     const float &l_grad,
+//     const float &l_eucl;
+
 
 torch::Tensor generalised_geodesic2d(
     torch::Tensor &image,
@@ -121,3 +136,18 @@ torch::Tensor generalised_geodesic3d_toivanen(
     const float &l_grad,
     const float &l_eucl,
     const int &iterations);
+
+torch::Tensor generalised_geodesic2d_fastmarch(
+    torch::Tensor &image,
+    const torch::Tensor &mask,
+    const float &v,
+    const float &l_grad,
+    const float &l_eucl);
+
+// torch::Tensor generalised_geodesic3d_fastmarch(
+//     torch::Tensor &image,
+//     const torch::Tensor &mask,
+//     const std::vector<float> &spacing,
+//     const float &v,
+//     const float &l_grad,
+//     const float &l_eucl);
