### 2022-08-14 3c1b486
adding working FastMarch2d with examples
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
