+++ FastGeodis/geodis_fastmarch.cpp
@@ -47,7 +47,7 @@ struct Point2D
 void insert_point_to_list(std::vector<Point2D> * list, int start_position,  Point2D p)
 {
     int insert_idx = list->size();
-    for(int i = start_position; i < list->size(); i++)
+    for(int i = start_position; i < int(list->size()); i++)
     {
         if(list->at(i).distance < p.distance)
         {
@@ -61,7 +61,7 @@ void insert_point_to_list(std::vector<Point2D> * list, int start_position,  Poin
 void update_point_in_list(std::vector<Point2D> * list, Point2D p)
 {
     int remove_idx = -1;
-    for(int i = 0; i < list->size(); i++)
+    for(int i = 0; i < int(list->size()); i++)
     {
         if(list->at(i).w == p.w && list->at(i).h == p.h)
         {
@@ -74,91 +74,91 @@ void update_point_in_list(std::vector<Point2D> * list, Point2D p)
 }
 
 void add_new_accepted_point(
-    const torch::Tensor &image,
-    torch::Tensor &distance,
-    torch::Tensor &state, 
+    torch::TensorAccessor<float, 4> image_ptr, 
+    torch::TensorAccessor<float, 4> distance_ptr, 
+    torch::TensorAccessor<signed char, 2> state_ptr, 
     std::vector<Point2D> * list, 
-    Point2D p, 
-    float l_grad,
-    float l_eucl)
+    const Point2D &p,
+    const int &channel, 
+    const int &height, 
+    const int &width, 
+    const float &l_grad,
+    const float &l_eucl)
 {
-    // batch, channel, height, width
-    const int channel = image.size(1);
-    const int height = image.size(2);
-    const int width = image.size(3);
-
-    auto image_ptr = image.accessor<float, 4>();
-    auto distance_ptr = distance.accessor<float, 4>();
-    auto state_ptr = state.accessor<int, 2>();
-
     int w = p.w;
     int h = p.h;
     int nh, nw, temp_state;
     float l_dist, space_dis, delta_dis, old_dis, new_dis;
     
-    for (int dh = -1; dh <= 1; dh++)
+    const int dh_f[9] = {-1, -1, -1,  0,  0,  0,   1,  1,  1};
+    const int dw_f[9] = {-1,  0,  1, -1,  0,  1,  -1,  0,  1};
+
+    const float local_dist_f[9] = {sqrt(2), float(1), sqrt(2), float(1), 0, float(1), sqrt(2), float(1), sqrt(2)};
+    
+    for(int ind = 0; ind < 9; ind++)
     {
-        for(int dw = -1; dw <= 1; dw++)
+        int dh = dh_f[ind];
+        int dw = dw_f[ind];
+
+        space_dis = local_dist_f[ind];
+
+        if(dh == 0 && dw == 0) 
         {
-            if(dh == 0 && dw == 0) 
+            continue;
+        }
+
+        nh = dh + h;
+        nw = dw + w;
+        
+        if(nh >=0 && nh < height && nw >=0 && nw < width)
+        {
+            temp_state = state_ptr[nh][nw];
+
+            if(temp_state == 0)
             {
                 continue;
             }
-
-            nh = dh + h;
-            nw = dw + w;
             
-            if(nh >=0 && nh < height && nw >=0 && nw < width)
+            l_dist = 0.0;
+            if (channel == 1)
             {
-                temp_state = state_ptr[nh][nw];
-
-                if(temp_state == 0)
+                l_dist = l1distance_fastmarch(
+                    image_ptr[0][0][h][w], 
+                    image_ptr[0][0][nh][nw]); 
+            }
+            else
+            {
+                for (int c_i=0; c_i < channel; c_i++)
                 {
-                    continue;
-                }
[…]
