+++ FastGeodis/geodis_fastmarch.cpp
@@ -271,185 +271,242 @@ torch::Tensor generalised_geodesic2d_fastmarch_cpu(
     return distance;
 }
 
-// void geodesic3d_fastmarch_cpu(
-//     const torch::Tensor &image,
-//     torch::Tensor &distance,
-//     const std::vector<float> &spacing,
-//     const float &l_grad,
-//     const float &l_eucl)
-// {
-//     // batch, channel, depth, height, width
-//     const int channel = image.size(1);
-//     const int depth = image.size(2);
-//     const int height = image.size(3);
-//     const int width = image.size(4);
+struct Point3D
+{
+    float distance;
+    int d;
+    int w;
+    int h;
+};
 
-//     auto image_ptr = image.accessor<float, 5>();
-//     auto distance_ptr = distance.accessor<float, 5>();
+void insert_point_to_list(std::vector<Point3D> * list, int start_position,  Point3D p)
+{
+    int insert_idx = list->size();
+    for(int i = start_position; i < int(list->size()); i++)
+    {
+        if(list->at(i).distance < p.distance)
+        {
+            insert_idx = i;
+            break;
+        }
+    }
+    list->insert(insert_idx + list->begin(), p);
+}
 
-//     // distances for forward
-//     const int dz_f[13] = {-1, -1, -1, -1, -1, 0, 0, 0, 0, 1, 1, 1, 1};
-//     const int dh_f[13] = {-1, -1, -1, 0, 0, -1, -1, -1, 0, -1, -1, -1, 0};
-//     const int dw_f[13] = {-1, 0, 1, -1, 0, -1, 0, 1, -1, -1, 0, 1, -1};
+void update_point_in_list(std::vector<Point3D> * list, Point3D p)
+{
+    int remove_idx = -1;
+    for(int i = 0; i < int(list->size()); i++)
+    {
+        if(list->at(i).d == p.d && list->at(i).h == p.h && list->at(i).w == p.w)
+        {
+            remove_idx = i;
+            break;
+        }
+    }
+    list->erase(remove_idx + list->begin());
+    insert_point_to_list(list, remove_idx, p);
+}
 
-//     float local_dist_f[13];
-//     for (int i = 0; i < 13; i++)
-//     {
-//         float ld = 0.0;
-//         if (dz_f[i] != 0)
-//         {
-//             ld += spacing[0] * spacing[0];
-//         }
+void add_new_accepted_point(
+    torch::TensorAccessor<float, 5> image_ptr, 
+    torch::TensorAccessor<float, 5> distance_ptr, 
+    torch::TensorAccessor<signed char, 3> state_ptr, 
+    std::vector<Point3D> * list,
+    Point3D p,
+    std::vector<float> spacing, 
+    int depth, 
+    int height, 
+    int width, 
+    int channel, 
+    const float &l_grad,
+    const float &l_eucl)
+{
+    int d = p.d;
+    int h = p.h;
+    int w = p.w;
+    float l_dist, delta_dis, old_dis, new_dis;
+    
+    for(int dd = -1; dd <= 1; dd++)
+    {
+        for(int dh = -1; dh <= 1; dh++)
+        {
+            for(int dw = -1; dw <= 1; dw++)
+            {
+                if(dd == 0 && dh == 0 && dw == 0) continue;
+                int nd = dd + d;
+                int nh = dh + h;
+                int nw = dw + w;
+                
+                if(nd >=0 && nd < depth && nh >=0 && nh < height && nw >=0 && nw < width )
+                {
+                    int temp_state = state_ptr[nd][nh][nw];
+                    if(temp_state == 0)
+                    {
+                        continue;
+                    }
 
-//         if (dh_f[i] != 0)
-//         {
[…]
