+++ FastGeodis/geodis_fastmarch.cpp
@@ -87,21 +87,29 @@ void add_new_accepted_point(
 {
     int w = p.w;
     int h = p.h;
-    int nh, nw, temp_state;
+    int dh, dw, nh, nw, temp_state;
     float l_dist, space_dis, delta_dis, old_dis, new_dis;
     
-    const int dh_f[9] = {-1, -1, -1,  0,  0,  0,   1,  1,  1};
-    const int dw_f[9] = {-1,  0,  1, -1,  0,  1,  -1,  0,  1};
+    constexpr int dh_f[9] = {
+        -1, -1, -1,  
+         0,  0,  0,   
+         1,  1,  1
+        };
+    constexpr int dw_f[9] = {
+        -1,  0,  1, 
+        -1,  0,  1,  
+        -1,  0,  1
+        };
 
-    const float local_dist_f[9] = {
+    constexpr float local_dist_f[9] = {
         sqrt(float(2.)), float(1), sqrt(float(2.)), 
         float(1.), float(0.), float(1.), 
         sqrt(float(2.)), float(1.), sqrt(float(2.))};
     
     for(int ind = 0; ind < 9; ind++)
     {
-        int dh = dh_f[ind];
-        int dw = dw_f[ind];
+        dh = dh_f[ind];
+        dw = dw_f[ind];
 
         space_dis = local_dist_f[ind];
 
@@ -313,84 +321,133 @@ void add_new_accepted_point(
     torch::TensorAccessor<float, 5> distance_ptr, 
     torch::TensorAccessor<signed char, 3> state_ptr, 
     std::vector<Point3D> * list,
-    Point3D p,
-    std::vector<float> spacing, 
-    int depth, 
-    int height, 
-    int width, 
-    int channel, 
+    const Point3D &p,
+    const std::vector<float> &spacing, 
+    const int &channel, 
+    const int &depth, 
+    const int &height, 
+    const int &width, 
     const float &l_grad,
     const float &l_eucl)
 {
     int d = p.d;
     int h = p.h;
     int w = p.w;
-    float l_dist, delta_dis, old_dis, new_dis;
+    int dd, dh, dw, nd, nh, nw, temp_state;
+    float l_dist, space_dis, delta_dis, old_dis, new_dis;
+
+    constexpr int dd_f[27] = { 
+        -1, -1, -1, -1, -1, -1, -1, -1, -1, 
+        0,  0,  0,  0,  0,  0,  0,  0,  0, 
+        1,  1,  1,  1,  1,  1,  1,  1,  1
+        };
+    constexpr int dh_f[27] = {
+        -1, -1, -1,  0,  0,  0,  1,  1,  1, 
+        -1, -1, -1,  0,  0,  0,  1,  1,  1,
+        -1, -1, -1,  0,  0,  0,  1,  1,  1
+        };
+    constexpr int dw_f[27] = { 
+        -1,  0,  1, -1,  0,  1, -1,  0,  1, 
+        -1,  0,  1, -1,  0,  1, -1,  0,  1, 
+        -1,  0,  1, -1,  0,  1, -1,  0,  1
+        };
     
-    for(int dd = -1; dd <= 1; dd++)
+    float local_dist_f[27];
+    for (int ind = 0; ind < 27; ind++)
     {
-        for(int dh = -1; dh <= 1; dh++)
+        float ld = 0.0;
+        if (dd_f[ind] != 0)
         {
-            for(int dw = -1; dw <= 1; dw++)
+            ld += spacing[0] * spacing[0];
+        }
+
+        if (dh_f[ind] != 0)
+        {
+            ld += spacing[1] * spacing[1];
+        }
+
+        if (dw_f[ind] != 0)
+        {
+            ld += spacing[2] * spacing[2];
[…]
