+++ samples/demo3d.py
@@ -11,7 +11,7 @@ import torch
 
 
 def demo_geodesic_distance3d(image_path, seed_pos):
-    SHOW_JOINT_HIST = False
+    SHOW_JOINT_HIST = True
     image_folder = os.path.dirname(image_path)
     image_sitk = sitk.ReadImage(image_path)
     input_image = sitk.GetArrayFromImage(image_sitk)
@@ -36,6 +36,8 @@ def demo_geodesic_distance3d(image_path, seed_pos):
         FastGeodis.generalised_geodesic3d_fastmarch(
             input_image_pt, seed_image_pt, spacing, 1e10, 1.0
         )
+        .cpu()
+        .numpy()
     )
     fastmarch_time = time.time() - tic
 
@@ -79,8 +81,8 @@ def demo_geodesic_distance3d(image_path, seed_pos):
         fastraster_time_gpu = time.time() - tic
 
     print(
-        "Toivanen's CPU raster: {:.6f} s \nFastGeodis CPU raster: {:.6f} s".format(
-            toivanenraster_time, fastraster_time_cpu
+        "Fast Marching CPU: {:.6} s \nToivanen's CPU raster: {:.6f} s \nFastGeodis CPU raster: {:.6f} s".format(
+            fastmarch_time, toivanenraster_time, fastraster_time_cpu
         )
     )
     if device:
@@ -104,6 +106,7 @@ def demo_geodesic_distance3d(image_path, seed_pos):
     input_image = np.asarray(input_image, np.uint8)
 
     image_slice = input_image[10]
+    fastmarch_output_slice = fastmarch_output[10]
     toivanenraster_output_slice = toivanenraster_output[10]
     fastraster_output_cpu_slice = fastraster_output_cpu[10]
     if device:
@@ -117,15 +120,20 @@ def demo_geodesic_distance3d(image_path, seed_pos):
     plt.axis("off")
     plt.title("(a) Input image")
 
+    plt.subplot(2, 4, 5)
+    plt.imshow(fastmarch_output_slice)
+    plt.axis("off")
+    plt.title("(b) Fast Marching (cpu) | ({:.4f} s)".format(fastmarch_time))
+
     plt.subplot(2, 4, 2)
     plt.imshow(toivanenraster_output_slice)
     plt.axis("off")
-    plt.title("(b) Toivanen's Raster (cpu) | ({:.4f} s)".format(toivanenraster_time))
+    plt.title("(c) Toivanen's Raster (cpu) | ({:.4f} s)".format(toivanenraster_time))
 
     plt.subplot(2, 4, 3)
     plt.imshow(fastraster_output_cpu_slice)
     plt.axis("off")
-    plt.title("(c) FastGeodis (cpu) | ({:.4f} s)".format(fastraster_time_cpu))
+    plt.title("(e) FastGeodis (cpu) | ({:.4f} s)".format(fastraster_time_cpu))
 
     plt.subplot(2, 4, 6)
     plt.imshow(toivanenraster_output_slice)
@@ -136,37 +144,37 @@ def demo_geodesic_distance3d(image_path, seed_pos):
         plt.subplot(2, 4, 7)
         plt.imshow(fastraster_output_gpu_slice)
         plt.axis("off")
-        plt.title("(e) FastGeodis (gpu) | ({:.4f} s)".format(fastraster_time_gpu))
+        plt.title("(f) FastGeodis (gpu) | ({:.4f} s)".format(fastraster_time_gpu))
 
     diff = (
-        abs(toivanenraster_output - fastraster_output_cpu)
-        / (toivanenraster_output + 1e-7)
+        abs(fastmarch_output - fastraster_output_cpu)
+        / (fastmarch_output + 1e-7)
         * 100
     )
-    diff_vol = toivanenraster_output - fastraster_output_cpu
[…]
