### 2022-08-15 d5d1897
optimise fast march execution
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
