### 2022-08-15 595db78
improve unit tests
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
