+++ tests/utils.py
@@ -1,0 +1,235 @@
+from functools import partial, wraps
+import unittest
+import torch
+
+try:
+    import FastGeodis
+except:
+    print(
+        "Unable to load FastGeodis for unittests\nMake sure to install using: python setup.py install"
+    )
+    exit()
+
+DEVICES_TO_RUN = ["cpu", "cuda"]
+CONF_2D_CPU = [("cpu", 2, bas) for bas in [16, 32, 64]]
+CONF_2D_CUDA = [("cuda", 2, bas) for bas in [16, 32, 64]]
+CONF_2D = CONF_2D_CPU + CONF_2D_CUDA
+
+CONF_3D_CPU = [("cpu", 3, bas) for bas in [16, 32, 64]]
+CONF_3D_CUDA = [("cuda", 3, bas) for bas in [16, 32, 64]]
+CONF_3D = CONF_3D_CPU + CONF_3D_CUDA
+CONF_ALL = CONF_2D + CONF_3D
+
+CONF_ALL_CPU = CONF_2D_CPU + CONF_3D_CPU
+
+# Fast March is compute intensive, so limit 2d and 3d cases to run with it
+CONF_2D_CPU_FM = [CONF_2D_CPU[0]]
+CONF_3D_CPU_FM = [CONF_3D_CPU[0]]
+CONF_ALL_CPU_FM = CONF_2D_CPU_FM + CONF_3D_CPU_FM
+
+
+
+def skip_if_no_cuda(obj):
+    return unittest.skipUnless(torch.cuda.is_available(), "Skipping CUDA-based tests")(
+        obj
+    )
+
+
+def run_cuda_if_available(fn):
+    @wraps(fn)
+    def wrapper(*args, **kwargs):
+        if args[1] == "cuda":
+            if torch.cuda.is_available():
+                return fn(*args, **kwargs)
+            else:
+                raise unittest.SkipTest("skipping as cuda device not found")
+        else:
+            return fn(*args, **kwargs)
+
+    return wrapper
+
+
+def fastgeodis_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
+    return FastGeodis.generalised_geodesic2d(image, softmask, v, lamb, iter)
+
+
+def fastgeodis_generalised_geodesic_distance_3d(
+    image, softmask, v, lamb, iter, spacing
+):
+    return FastGeodis.generalised_geodesic3d(image, softmask, spacing, v, lamb, iter)
+
+
+def fastgeodis_signed_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
+    return FastGeodis.signed_generalised_geodesic2d(image, softmask, v, lamb, iter)
+
+
+def fastgeodis_signed_generalised_geodesic_distance_3d(
+    image, softmask, v, lamb, iter, spacing
+):
+    return FastGeodis.signed_generalised_geodesic3d(
+        image, softmask, spacing, v, lamb, iter
+    )
+
+
+def toivanen_signed_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
+    return FastGeodis.signed_generalised_geodesic2d_toivanen(
+        image, softmask, v, lamb, iter
+    )
+
+
+def toivanen_signed_generalised_geodesic_distance_3d(
+    image, softmask, v, lamb, iter, spacing
+):
+    return FastGeodis.signed_generalised_geodesic3d_toivanen(
+        image, softmask, spacing, v, lamb, iter
+    )
+
+
+def fastmarch_signed_generalised_geodesic_distance_2d(image, softmask, v, lamb, iter):
+    return FastGeodis.signed_generalised_geodesic2d_fastmarch(image, softmask, v, lamb)
+
+
[…]
