### 2022-08-15 b27cbf5
update 3d fast march demos
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
