+++ samples/simpledemo3d_signed.py
@@ -92,3 +92,37 @@ plt.imshow(euclidean_dist_toivanen[10])
 plt.plot(seed_pos[2], seed_pos[1], "ro")
 
 plt.show()
+
+
+# Fast Marching Method
+
+# Fast Marching method only support CPU
+image_pt = image_pt.to("cpu")
+mask_pt = mask_pt.to("cpu")
+
+lamb = 1.0  # <-- Geodesic distance transform
+geodesic_dist_fastmarch = FastGeodis.signed_generalised_geodesic3d_fastmarch(
+    image_pt, mask_pt, spacing, v, lamb
+)
+geodesic_dist_fastmarch = np.squeeze(geodesic_dist_fastmarch.cpu().numpy())
+
+lamb = 0.0  # <-- Euclidean distance transform
+euclidean_dist_fastmarch = FastGeodis.signed_generalised_geodesic3d_fastmarch(
+    image_pt, mask_pt, spacing, v, lamb
+)
+euclidean_dist_fastmarch = np.squeeze(euclidean_dist_fastmarch.cpu().numpy())
+
+plt.figure(figsize=(12, 4))
+plt.subplot(1, 3, 1)
+plt.imshow(image[10])
+plt.plot(seed_pos[2], seed_pos[1], "ro")
+
+plt.subplot(1, 3, 2)
+plt.imshow(geodesic_dist_fastmarch[10])
+plt.plot(seed_pos[2], seed_pos[1], "ro")
+
+plt.subplot(1, 3, 3)
+plt.imshow(euclidean_dist_fastmarch[10])
+plt.plot(seed_pos[2], seed_pos[1], "ro")
+
+plt.show()
