### 2022-08-15 518a175
update docs with addition of Fast March description
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
