+++ docs/source/refs.bib
@@ -161,3 +161,14 @@
             Xiaoqiang~Zheng},
   year   = {2015}
 }
+
+@article{sethian1999fast,
+  title={Fast marching methods},
+  author={Sethian, James A},
+  journal={SIAM review},
+  volume={41},
+  number={2},
+  pages={199--235},
+  year={1999},
+  publisher={SIAM}
+}
