### 2022-08-15 #4532 @masadcv
Just a brief update, I have now included a third method (https://github.com/masadcv/FastGeodis/pull/26) based on [*]. This serves as golden reference for checking accuracy of the implemented Fast Geodis methods.

[*] Sethian, James A. "Fast marching methods." SIAM review 41.2 (1999): 199-235.
