### 2022-08-19 #4532 @masadcv
Hi @vissarion ,
We are wondering if there is any update on the scope query for this submission? If so, can you please guide us on the next steps/improvements needed from our end? Many thanks!
