### 2022-08-21 72bcb8c
update with multi-thread signed distance calculation
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
