+++ FastGeodis/fastgeodis.cpp
@@ -30,6 +30,7 @@
 
 #include <torch/extension.h>
 #include <iostream>
+#include <future>
 #include <vector>
 #include "fastgeodis.h"
 #include "common.h"
@@ -40,7 +41,7 @@
 
 #define VERBOSE 0
 
-torch::Tensor generalised_geodesic2d(torch::Tensor &image, const torch::Tensor &mask, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
+torch::Tensor generalised_geodesic2d(const torch::Tensor &image, const torch::Tensor &mask, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
 {
     #if VERBOSE
         #ifdef _OPENMP
@@ -88,7 +89,7 @@ torch::Tensor generalised_geodesic2d(torch::Tensor &image, const torch::Tensor &
     return generalised_geodesic2d_cpu(image, mask, v, l_grad, l_eucl, iterations);
 }
 
-torch::Tensor generalised_geodesic3d(torch::Tensor &image, const torch::Tensor &mask, const std::vector<float> &spacing, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
+torch::Tensor generalised_geodesic3d(const torch::Tensor &image, const torch::Tensor &mask, const std::vector<float> &spacing, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
 {
     #if VERBOSE
         #ifdef _OPENMP
@@ -144,7 +145,7 @@ torch::Tensor generalised_geodesic3d(torch::Tensor &image, const torch::Tensor &
     return generalised_geodesic3d_cpu(image, mask, spacing, v, l_grad, l_eucl, iterations);
 }
 
-torch::Tensor generalised_geodesic2d_toivanen(torch::Tensor &image, const torch::Tensor &mask, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
+torch::Tensor generalised_geodesic2d_toivanen(const torch::Tensor &image, const torch::Tensor &mask, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
 {
 
     // check input dimensions
@@ -157,7 +158,7 @@ torch::Tensor generalised_geodesic2d_toivanen(torch::Tensor &image, const torch:
     return generalised_geodesic2d_toivanen_cpu(image, mask, v, l_grad, l_eucl, iterations);
 }
 
-torch::Tensor generalised_geodesic3d_toivanen(torch::Tensor &image, const torch::Tensor &mask, const std::vector<float> &spacing, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
+torch::Tensor generalised_geodesic3d_toivanen(const torch::Tensor &image, const torch::Tensor &mask, const std::vector<float> &spacing, const float &v, const float &l_grad, const float &l_eucl, const int &iterations)
 {
     // check input dimensions
     check_input_dimensions(image, mask, 5);
@@ -175,7 +176,7 @@ torch::Tensor generalised_geodesic3d_toivanen(torch::Tensor &image, const torch:
     return generalised_geodesic3d_toivanen_cpu(image, mask, spacing, v, l_grad, l_eucl, iterations);
 }
 
-torch::Tensor generalised_geodesic2d_fastmarch(torch::Tensor &image, const torch::Tensor &mask, const float &v, const float &l_grad, const float &l_eucl)
[…]
