+++ FastGeodis/geodis_fastmarch.cpp
@@ -275,7 +275,7 @@ void geodesic2d_fastmarch_cpu(
 }
 
 torch::Tensor generalised_geodesic2d_fastmarch_cpu(
-    torch::Tensor &image,
+    const torch::Tensor &image,
     const torch::Tensor &mask,
     const float &v,
     const float &l_grad,
@@ -573,7 +573,7 @@ void geodesic3d_fastmarch_cpu(
 }
 
 torch::Tensor generalised_geodesic3d_fastmarch_cpu(
-    torch::Tensor &image,
+    const torch::Tensor &image,
     const torch::Tensor &mask,
     const std::vector<float> &spacing,
     const float &v,
