### 2022-08-29 #4532 @jacobmerson
@vissarion thanks for doing the scope query. @masadcv let me know when you have an updated version of the paper/docs and I'll try to get everything finalized.
