### 2022-09-01 #4532 @masadcv
Many thanks @vissarion for confirming the scope.

@jacobmerson as request, I have updated docs/paper. Please find a copy of the updated paper above this comment.
I look forward to receiving your finalized feedback. Many thanks!
