+++ paper/paper.md
@@ -27,10 +27,10 @@ bibliography: paper.bib
 # Summary 
 
   
-Geodesic and Euclidean distance transforms have been widely used in a number of applications, where distance from a set of reference points is computed. Methods from recent years have shown effectiveness in applying Geodesic distance transform to interactively annotate 3D medical imaging data [@wang2018deepigeos; @criminisi2008geos], where it enables providing segmentation labels, i.e. voxel-wise labels, for different objects of interests. Despite existing methods for efficient computation of Geodesic distance transform on GPU and CPU devices [@criminisiinteractive; @criminisi2008geos; @weber2008parallel; @toivanen1996new], an open-source implementation of such methods does not exist. 
-On the contrary, efficient methods for Euclidean distance transform [@felzenszwalb2012distance] have open-source implementations [@tensorflow2015-whitepaper; @eucildeantdimpl]. Existing libraries, e.g. [@geodistk], provide C++ implementations of Geodesic distance transform, however they lack efficient utilisation of the underlying hardware and hence result in significant computation time especially when applying them on 3D medical imaging volumes.  
+Geodesic and Euclidean distance transforms have been widely used in a number of applications where distance from a set of reference points is computed. Methods from recent years have shown effectiveness in applying Geodesic distance transform to interactively annotate 3D medical imaging data [@wang2018deepigeos; @criminisi2008geos]. The Geodesic distance transform enables providing segmentation labels, i.e., voxel-wise labels, for different objects of interests. Despite existing methods for efficient computation of the Geodesic distance transform on GPU and CPU devices [@criminisiinteractive; @criminisi2008geos; @weber2008parallel; @toivanen1996new], an open-source implementation of such methods on the GPU does not exist. 
+On the contrary, efficient methods for the computation of the Euclidean distance transform [@felzenszwalb2012distance] have open-source implementations [@tensorflow2015-whitepaper; @eucildeantdimpl]. Existing libraries, e.g., [@geodistk], provide C++ implementations of the Geodesic distance transform, however they lack efficient utilisation of the underlying hardware and hence result in significant computation time especially when applying them on 3D medical imaging volumes.  
 
-The `FastGeodis` package provides an efficient implementation for computing Geodesic and Euclidean distance transforms (or a mixture of both) targeting efficient utilisation of CPU and GPU hardwares. In particular, it implements paralellisable raster scan method from [@criminisiinteractive], where elements in row (2D) or plane (3D) can be computed with parallel threads. This package is able to handle 2D as well as 3D data where it achieves up to 20x speed-up on CPU and up to 74x speed-up on GPU as compared to existing open-source libraries [@geodistk], which uses non-parallelisable single-thread CPU implementation. The performance speed-ups reported here were evaluated using 3D volume data on Nvidia GeForce Titan X (12 GB) with 6-Core Intel Xeon E5-1650 CPU. Further in-depth comparison of performance improvements is discussed in `FastGeodis` \href{https://masad.ml/FastGeodis/}{documentation}. 
[…]
