### 2022-09-06 a9629c5
fix docs link in paper
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
