### 2022-09-06 a5c5d7e
adding patent references
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
