+++ paper/paper.bib
@@ -149,3 +149,30 @@
             Xiaoqiang~Zheng},
   year   = {2015}
 }
+
+@misc{bronstein2013parallel,
+  title={Parallel approximation of distance maps},
+  author={Bronstein, Alexander and Bronstein, Michael and Devir, Yohai and Weber, Ofir and Kimmel, Ron},
+  year={2013},
+  month=feb # "~12",
+  publisher={Google Patents},
+  note={US Patent 8,373,716}
+}
+
+@misc{bronstein2015parallel,
+  title={Parallel approximation of distance maps},
+  author={Bronstein, Alexander and Bronstein, Michael and Kimmel, Ron and Devir, Yohai and Weber, Ofir},
+  year={2015},
+  month=mar # "~17",
+  publisher={Google Patents},
+  note={US Patent 8,982,142}
+}
+
+@misc{bronstein2016parallel,
+  title={Parallel approximation of distance maps},
+  author={Bronstein, Alexander and Bronstein, Michael and Kimmel, Ron and Devir, Yohai and Weber, Ofir},
+  year={2016},
+  month=nov # "~8",
+  publisher={Google Patents},
+  note={US Patent 9,489,708}
+}
