### 2022-09-06 #4532 @masadcv
Many thanks @jacobmerson !
These are extremely helpful comments for improving the article further.

I have resolved all issues mentioned, including merging your PR for grammar fixes and adding two separate PRs to address the links and missing patents citations.

The patents don't usually have a DOI link, hence I am unable to include DOI for the patent citations. I hope that is okay with JOSS publication guidelines.
