### 2022-10-17 #4532 @vissarion
@masadcv I am searching for one more (new) reviewer while waiting for @yhtang to respond.

Hi @dataplayer12, @yogi-tud, @robertszafa, @kris-rowe, @zhampel as experts in C++/CUDA could you please help us by reviewing this submission for JOSS?
