### 2022-10-18 #4532 @vissarion
Hi @dataplayer12 thanks for accepting to help by reviewing this submission!

We aim for reviews to be completed within about 2-4 weeks. Please let me know if you require some more time.

Also, please find below links for the review guidelines, criteria and checklist.

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
https://joss.readthedocs.io/en/latest/review_criteria.html
https://joss.readthedocs.io/en/latest/review_checklist.html
