### 2022-11-01 #4532 @vissarion
@dataplayer12 and @zhampel I am copying / gathering here some information from the thread above just to be sure we are all on the some page.

Each of you will have a separate checklist that you should update when carrying out your review.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if you require some more time.

Also, please find below links for the review guidelines, criteria and checklist.
 https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
 https://joss.readthedocs.io/en/latest/review_criteria.html
 https://joss.readthedocs.io/en/latest/review_checklist.html
