### 2022-11-03 #4532 @dataplayer12
Hi, sorry for taking longer than anticipated.
* Having read through the paper, the code and other reviewer's comments, **I recommend accepting this manuscript**.

* As @yhtang pointed out in particular, the CUDA code isn't the highest quality and does leave performance on the table. However, I believe it would be unfair to compare the author's implementation against some theoretically optimal solution. Rather we should compare it against currently existing open solutions. On this metric, the FastGeodis codebase does well, and the claimed performance is borne out in testing.
* PyTorch bindings are excellent and should open up this work to the ML community.
* Like others have pointed out, I too, recommend that the authors add more mathematical description of the Geodesic distance transform and the algorithm from Criminisi et. al., 2009.
