### 2022-11-15 #4532 @masadcv
Many thanks @dataplayer12 for your detailed review and for recommendation to accept.

- Regarding mathematical description, according to JOSS guidelines, we are not allowed to include mathematical description / methodology in paper (as per editor comment here: https://github.com/openjournals/joss-reviews/issues/4532#issuecomment-1195144324).  We have already added mathematical description in documentation at: https://fastgeodis.readthedocs.io/en/latest/methodology.html - can you check if this is sufficient, if not we can work on expanding it with more details.

@vissarion given we have more reviews now, what are the next steps for this submission?
