### 2022-11-22 af5ae76
release v1.0.0 for Zenodo, JOSS paper
Signed-off-by: masadcv <muhammad.asad@kcl.ac.uk>
