### 2022-11-22 #4532 @masadcv
@vissarion many thanks for instructions on the next step.
I have just released v1.0.0 with all changes from this review.

This has been archived at Zenodo at the following doi:

[![DOI](https://zenodo.org/badge/375526483.svg)](https://zenodo.org/badge/latestdoi/375526483)

Please let me know if there is anything else needed from our end at this stage.
