+++ paper/paper.bib
@@ -42,11 +42,11 @@
 
 @misc{eucildeantdimpl,
   author    = {Seung-Lab, },
-  title     = {Multi-Label Anisotropic 3D Euclidean Distance Transform (MLAEDT-3D)},
+  title     = {Multi-Label Anisotropic 3D {E}uclidean Distance Transform (MLAEDT-3D)},
   year      = {2018},
   publisher = {GitHub},
   journal   = {GitHub repository},
-  url       = {https://github.com/seung-lab/euclidean-distance-transform-3d}
+  url       = {https://github.com/seung-lab/dean-distance-transform-3d}
 }
 
 
@@ -70,7 +70,7 @@
 
 @inproceedings{criminisiinteractive,
   author    = {Criminisi, Antonio and Sharp, Toby and Siddiqui, Khan},
-  title     = {Interactive Geodesic Segmentation of n-Dimensional Medical Images on the Graphics Processor},
+  title     = {Interactive {G}eodesic Segmentation of n-Dimensional Medical Images on the Graphics Processor},
   booktitle = {Radiological Society of North America (RSNA)},
   year      = {2009},
   month     = {December},
@@ -151,7 +151,7 @@
 }
 
 @misc{bronstein2013parallel,
-  title={Parallel approximation of distance maps},
+  title={Parallel approximation of distance maps (US Patent 8,373,716)},
   author={Bronstein, Alexander and Bronstein, Michael and Devir, Yohai and Weber, Ofir and Kimmel, Ron},
   year={2013},
   month=feb # "~12",
@@ -160,7 +160,7 @@
 }
 
 @misc{bronstein2015parallel,
-  title={Parallel approximation of distance maps},
+  title={Parallel approximation of distance maps (US Patent 8,982,142)},
   author={Bronstein, Alexander and Bronstein, Michael and Kimmel, Ron and Devir, Yohai and Weber, Ofir},
   year={2015},
   month=mar # "~17",
@@ -169,7 +169,7 @@
 }
 
 @misc{bronstein2016parallel,
-  title={Parallel approximation of distance maps},
+  title={Parallel approximation of distance maps (US Patent 9,489,708)},
   author={Bronstein, Alexander and Bronstein, Michael and Kimmel, Ron and Devir, Yohai and Weber, Ofir},
   year={2016},
   month=nov # "~8",
