### 2022-11-23 #4532 @danielskatz
👋 @masadcv - I'm the track editor who is taking over at this point. I've proofread the paper, and have made a bunch of small suggestions in https://github.com/masadcv/FastGeodis/pull/36 - please merge this, or let me know what you disagree with, then we can continue the acceptance and publishing process
