+++ paper/paper.bib
@@ -46,7 +46,7 @@
   year      = {2018},
   publisher = {GitHub},
   journal   = {GitHub repository},
-  url       = {https://github.com/seung-lab/dean-distance-transform-3d}
+  url       = {https://github.com/seung-lab/euclidean-distance-transform-3d}
 }
 
 
